### 2023-12-01 6e43660
added prior test with one observation
