+++ linfa/tests/test_discr_10_onlydiscr_rep_meas.py
@@ -1,0 +1,121 @@
+from linfa.run_experiment import experiment
+from linfa.transform import Transformation
+from linfa.discrepancy import Discrepancy
+import torch
+import random
+import numpy as np
+
+# Import rcr model
+from linfa.models.discrepancy_models import PhysChem
+
+def run_test():
+
+    exp = experiment()
+    exp.name = "only_discr_ex"
+    exp.flow_type           = 'maf'         # str: Type of flow (default 'realnvp')
+    exp.n_blocks            = 15            # int: Number of hidden layers   
+    exp.hidden_size         = 100           # int: Hidden layer size for MADE in each layer (default 100)
+    exp.n_hidden            = 1             # int: Number of hidden layers in each MADE
+    exp.activation_fn       = 'relu'        # str: Activation function used (default 'relu')
+    exp.input_order         = 'sequential'  # str: Input oder for create_mask (default 'sequential')
+    exp.batch_norm_order    = True          # bool: Order to decide if batch_norm is used (default True)
+    exp.save_interval       = 1000          # int: How often to sample from normalizing flow
+    
+    exp.input_size          = 2             # int: Dimensionalty of input (default 2)
+    exp.batch_size          = 200           # int: Number of samples generated (default 100)
+    exp.true_data_num       = 2             # double: Number of true model evaluted (default 2)
+    exp.n_iter              = 2501          # int: Number of iterations (default 25001)
+    exp.lr                  = 0.001         # float: Learning rate (default 0.003)
+    exp.lr_decay            = 0.9999        # float:  Learning rate decay (default 0.9999)
+    exp.log_interal         = 10            # int: How often to show loss stat (default 10)
+
+    exp.run_nofas           = True          # normalizing flow with adaptive surrogate
+    exp.surrogate_type      = 'discrepancy' # type of surrogate we are using
+    exp.surr_pre_it         = 10000         # int: Number of pre-training iterations for surrogate model
+    exp.surr_upd_it         = 1000          # int: Number of iterations for the surrogate model update
+
+    exp.annealing           = False
+    exp.calibrate_interval  = 300           # int: How often to update the surrogate model (default 1000)
+    exp.budget              = 216           # int: Total number of true model evaulations
+    exp.surr_folder         = "./" 
+    exp.use_new_surr        = True
+
+    exp.output_dir = './results/' + exp.name
+    exp.log_file = 'log.txt'
+    
+    exp.seed = random.randint(0, 10 ** 9)   # int: Random seed used
+    exp.n_sample = 5000                     # int: Batch size to generate final results/plots
+    exp.no_cuda = True                      # Running on CPU by default but teste on CUDA
+
+    exp.optimizer = 'RMSprop'
+    exp.lr_scheduler = 'ExponentialLR'
+
+    exp.device = torch.device('cuda:0' if torch.cuda.is_available() and not exp.no_cuda else 'cpu')
+
+    # Add temperatures and pressures for each evaluation
+    variable_inputs = [[350.0, 400.0, 450.0],
+                       [1.0, 2.0, 3.0, 4.0, 5.0]]
+
+    # Define model
+    langmuir_model = PhysChem(variable_inputs)
+    # Assign as experiment model
+    exp.model = langmuir_model
+    # Read data
+    exp.model.data = np.loadtxt('observations.csv', delimiter = ',', skiprows = 1)
+
+    if(len(exp.model.data.shape) < 2):
+        exp.model.data = np.expand_dims(exp.model.data, axis=0)
+    
+    # Form tensors for variables and results in observations
[…]
