+++ linfa/tests/test_discr_11_LF_nodiscr_LFData_TP1 copy_rep_meas.py
@@ -1,6 +1,5 @@
 from linfa.run_experiment import experiment
 from linfa.transform import Transformation
-from linfa.discrepancy import Discrepancy
 import torch
 import random
 import numpy as np
@@ -11,7 +10,7 @@ from linfa.models.discrepancy_models import PhysChem
 def run_test():
 
     exp = experiment()
-    exp.name = "only_discr_ex_rep_meas"
+    exp.name = "lf_no_disc_lf_data_tp1_rep_meas"
     exp.flow_type           = 'maf'         # str: Type of flow (default 'realnvp')
     exp.n_blocks            = 15            # int: Number of hidden layers   
     exp.hidden_size         = 100           # int: Hidden layer size for MADE in each layer (default 100)
@@ -24,14 +23,14 @@ def run_test():
     exp.input_size          = 2             # int: Dimensionalty of input (default 2)
     exp.batch_size          = 200           # int: Number of samples generated (default 100)
     exp.true_data_num       = 2             # double: Number of true model evaluted (default 2)
-    exp.n_iter              = 2501          # int: Number of iterations (default 25001)
+    exp.n_iter              = 25001         # int: Number of iterations (default 25001)
     exp.lr                  = 0.001         # float: Learning rate (default 0.003)
     exp.lr_decay            = 0.9999        # float:  Learning rate decay (default 0.9999)
     exp.log_interal         = 10            # int: How often to show loss stat (default 10)
 
-    exp.run_nofas           = True          # normalizing flow with adaptive surrogate
+    exp.run_nofas           = False         # normalizing flow with adaptive surrogate
     exp.surrogate_type      = 'discrepancy' # type of surrogate we are using
-    exp.surr_pre_it         = 1000         # int: Number of pre-training iterations for surrogate model
+    exp.surr_pre_it         = 1000          # int: Number of pre-training iterations for surrogate model
     exp.surr_upd_it         = 1000          # int: Number of iterations for the surrogate model update
 
     exp.annealing           = False
@@ -52,65 +51,101 @@ def run_test():
 
     exp.device = torch.device('cuda:0' if torch.cuda.is_available() and not exp.no_cuda else 'cpu')
 
+    # Define transformation
+    trsf_info = [['linear', -4.0, 4.0, 500.0, 1500.0],
+                 ['linear', -4.0, 4.0, -30000.0, -15000.0]]
+    trsf = Transformation(trsf_info)
+    
+    # Apply the transformation
+    exp.transform = trsf
+
     # Add temperatures and pressures for each evaluation
-    variable_inputs = [[350.0, 400.0, 450.0],
-                       [1.0, 2.0, 3.0, 4.0, 5.0]]
+    variable_inputs = [[350.0],
+                       [1.0]]
 
     # Define model
     langmuir_model = PhysChem(variable_inputs)
     # Assign as experiment model
     exp.model = langmuir_model
+
     # Read data
     exp.model.data = np.loadtxt('observations.csv', delimiter = ',', skiprows = 1)
 
     if(len(exp.model.data.shape) < 2):
         exp.model.data = np.expand_dims(exp.model.data, axis=0)
     
-    # Form tensors for variables and results in observations
-    var_grid_in = torch.tensor(exp.model.data[:,:2])
-    var_grid_out = torch.tensor(exp.model.data[:,2:])
+    # No surrogate
+    exp.surrogate = None
 
-    # Define surrogate
-    if(exp.run_nofas):
+    # Define log density
+    def log_density(calib_inputs, model, surrogate, transform):
+        
+        # Compute transformation by log Jacobian
+        adjust = transform.compute_log_jacob_func(calib_inputs)
 
[…]
