+++ linfa/tests/test_discr_11_LF_nodiscr_LFData_TP1_rep_meas.py
@@ -139,7 +139,7 @@ def generate_data():
     model = PhysChem(var_grid)
     
     # Generate data
-    model.genDataFile(use_true_model=False,num_observations=1)
+    model.genDataFile(use_true_model=False,num_observations=2)
 
 # Main code
 if __name__ == "__main__":

[…]
