### 2023-12-05 cc464ad
started script to plot discrepancy
