### 2023-12-08 67e491b
changes to print the disrepancy function on a user defined grid of TP
