+++ linfa/tests/test_discr_06_LF_withdiscr_HFData_TP1.py
@@ -84,6 +84,7 @@ def run_test():
 
         # Create new discrepancy
         exp.surrogate = Discrepancy(model_name=exp.name, 
+                                    model_folder=exp.output_dir,
                                     lf_model=exp.model.solve_t,
                                     input_size=exp.model.var_in.size(1),
                                     output_size=1,
