### 2023-12-08 1ab8525
upated plot disc, resolved prior error for TP1 cases
