### 2023-12-08 3bea243
implemented plotting functionalities for discrepancy
