+++ linfa/plot_disc.py
@@ -3,35 +3,83 @@ import matplotlib.pyplot as plt
 from linfa.discrepancy import Discrepancy
 import numpy as np
 import os
+import argparse
 
-# directory = 'results'
-# iteration = '25000'
-# experiment = 'test_lf_with_disc_hf_data_prior_TP1'
-# filename1 = 'outputs_lf'
-# filename2 = 'outputs_lf+discr'pi
-# filename3 = 'outputs_lf+discr+noise'
+def scale_limits(min,max,factor):
+    if(min>max):
+        temp = max
+        max = min
+        min = temp
+    center = 0.5*(min+max)
+    range = max - min
+    return center - factor*0.5*range, center + factor*0.5*range
 
-# path1 = os.path.join(directory, experiment, experiment + '_' + filename1 + '_' + iteration)
-# path2 = os.path.join(directory, experiment, experiment + '_' + filename2 + '_' + iteration)
-# path3 = os.path.join(directory, experiment, experiment + '_' + filename3 + '_' + iteration)
+def plot_disr_histograms(lf_file,lf_dicr_file,lf_discr_noise_file):
 
-# lf_model = np.loadtxt(path1)
-# lf_model_plus_disc = np.loadtxt(path2)
-# lf_model_plus_disc_plus_noise = np.loadtxt(path3)
+    # Read result files
+    lf_model = np.loadtxt(lf_file)
+    lf_model_plus_disc = np.loadtxt(lf_dicr_file)
+    lf_model_plus_disc_plus_noise = np.loadtxt(lf_discr_noise_file)
 
-<<<<<<< Updated upstream
-# plt.hist(lf_model[:,0], label = 'LF')
-# plt.hist(lf_model_plus_disc[:,0],  label = 'LF + disc')
-# plt.hist(lf_model_plus_disc_plus_noise[:,0],  label = 'LF + disc + noise')
-# plt.xlabel('Coverage')
-# plt.ylabel('Frequency')
-# plt.legend()
-# plt.show()
+    # Plot histograms
+    plt.hist(lf_model, label = 'LF', alpha =0.5)
+    plt.hist(lf_model_plus_disc,  label = 'LF + disc', alpha =0.5)
+    plt.hist(lf_model_plus_disc_plus_noise,  label = 'LF + disc + noise', alpha = 0.5)
+    plt.xlabel('Coverage')
+    plt.ylabel('Frequency')
+    plt.legend()
+    plt.show()
 
-# shape: number of ___?___ (15) x Batch size (5000)
-# print(np.shape(lf_model))
+def plot_discr_surface_2d(file_path,data_file,num_1d_grid_points,data_limit_factor):
 
-def plot_discrepancy(file_path,train_grid_in,train_grid_out,test_grid):
+    exp_name = os.path.basename(file_path)
+    dir_name = os.path.dirname(file_path) 
+
+    print(exp_name)
+    print(dir_name)
+
+    # Create new discrepancy
+    dicr = Discrepancy(model_name=exp_name, 
+                       model_folder=dir_name,
+                       lf_model=None,
+                       input_size=None,
+                       output_size=None,
+                       var_grid_in=None,
+                       var_grid_out=None)
+    dicr.surrogate_load()
+
+    # Get the number of dimensions for the aux variable
+    num_dim = dicr.var_grid_in.size(1)
+    if(num_dim == 2):
+        min_dim_1 = torch.min(dicr.var_grid_in[:,0])
+        max_dim_1 = torch.max(dicr.var_grid_in[:,0])
+        min_dim_2 = torch.min(dicr.var_grid_in[:,1])
[…]
