### 2023-12-09 242a8ae
Update run_experiment.py
Added -inf check into annealing
