### 2023-12-11 370b9b7
updated script to plot true process posterior across multiple temperatures and pressures and deleted unneeded comment on test 9
