### 2023-12-11 6b30d53
confirmed that histograms work for TP1 with repeated observations
