+++ linfa/tests/test_plot_discr.sh
@@ -1,6 +1,6 @@
  # discr_surface
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_lf_with_disc_hf_data_prior_TP15_rep_meas \
+							--name test_lf_with_disc_hf_data_prior_TP1_rep_meas \
 							--iter 2000 \
 							--mode histograms \
 							--num_points 10 \
