### 2023-12-12 4d01629
updated test scripts to save surrogate to experiment results, added save features to plot_disc, added axis labels to plot_disc
