+++ linfa/tests/test_plot_discr.sh
@@ -1,16 +1,16 @@
  # discr_surface
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_lf_with_disc_hf_data_prior_TP1_rep_meas \
-							--iter 2000 \
+							--name test_lf_with_disc_hf_data_prior_TP15_rep_meas \
+							--iter 15000 \
 							--mode histograms \
 							--num_points 10 \
 							--limfactor 1.0 \
 
 
 
-# python3 -m linfa.plot_disc --folder results/ \
-# 							--name test_lf_with_disc_hf_data_prior_TP15 \
-# 							--iter 2000 \
-# 							--mode discr_surface \
-# 							--num_points 40 \
-# 							--limfactor 2.0 \
+python3 -m linfa.plot_disc --folder results/ \
+							--name test_lf_with_disc_hf_data_prior_TP15_rep_meas \
+							--iter 15000 \
+							--mode discr_surface \
+							--num_points 40 \
+							--limfactor 2.0 \
