### 2023-12-12 d799ae1
made 1D histograms pub quality
