### 2023-12-12 ad65af2
made 2d histograms cvd friendly, added color to triangle surf plot, made both plots publication qual
