+++ linfa/plot_disc.py
@@ -8,13 +8,13 @@ from numpy import random
 from matplotlib.ticker import AutoMinorLocator, FormatStrFormatter
 
 def scale_limits(min,max,factor):
-    if(min>max):
+    if(min > max):
         temp = max
         max = min
         min = temp
-    center = 0.5*(min+max)
+    center = 0.5 * (min + max)
     range = max - min
-    return center - factor*0.5*range, center + factor*0.5*range
+    return center - factor * 0.5 * range, center + factor * 0.5 * range
 
 def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, out_dir, sample_size = 250):
 
@@ -31,17 +31,18 @@ def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, 
     if num_dim == 1:
         # Plot histograms
         plt.figure(figsize = (6,4))
+        ax = plt.gca()
         plt.hist(lf_model,  label = r'$\eta \vert \mathbf{\theta}$', alpha = 0.5, density = True, hatch = '/')
         plt.hist(lf_model_plus_disc,  label = r'$\zeta \vert \mathbf{\theta}, \delta$', alpha = 0.5, density = True)
         plt.hist(lf_model_plus_disc_plus_noise,  label = r'$y \vert \mathbf{\theta}, \delta, \epsilon$', alpha = 0.5, density = True, hatch = '.')
-        plt.xlabel('Coverage', fontweight = 'bold', fontsize = 16)
-        plt.ylabel('Density', fontweight = 'bold', fontsize = 16)
         plt.axvline(data[2], label = r'$y$', color = 'k', linewidth = 3)
         plt.legend(fontsize = 14)
-        plt.gca().xaxis.set_minor_locator(AutoMinorLocator())
-        plt.gca().yaxis.set_minor_locator(AutoMinorLocator())
-        plt.gca().xaxis.set_major_formatter(FormatStrFormatter("%.2f"))
-        plt.gca().tick_params(axis = 'both', which = 'both', direction = 'in', top = True, right = True, labelsize = 15)
+        ax.set_xlabel('Coverage [ ]', fontweight = 'bold', fontsize = 16)
+        ax.set_ylabel('Density', fontweight = 'bold', fontsize = 16)
+        ax.xaxis.set_minor_locator(AutoMinorLocator())
+        ax.yaxis.set_minor_locator(AutoMinorLocator())
+        ax.xaxis.set_major_formatter(FormatStrFormatter("%.2f"))
+        ax.tick_params(axis = 'both', which = 'both', direction = 'in', top = True, right = True, labelsize = 15)
         plt.tight_layout()
         plt.savefig(out_dir+'hist.png', bbox_inches = 'tight', dpi = 300)
         plt.close()
@@ -51,6 +52,8 @@ def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, 
         batch_size = len(lf_model_plus_disc[0])
         temps = np.unique(data[:, 0])
         pressures = np.unique(data[:, 1])
+        no_reps = len(data[0,2:])
+        observed = data[:,2:].reshape(len(temps), len(pressures), no_reps)
 
          ## Prepare for random sampling of batches
         lf_model_plus_disc = lf_model_plus_disc.reshape(len(temps), len(pressures), batch_size)
@@ -58,8 +61,10 @@ def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, 
         sample = np.zeros([len(temps), len(pressures), sample_size]) # Initialize
 
          # For plotting
-        plt.figure(figsize = (4,4))
[…]
