+++ linfa/tests/test_plot_discr.sh
@@ -1,6 +1,6 @@
  # discr_surface
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_lf_with_disc_hf_data_prior_TP1 \
+							--name test_lf_with_disc_hf_data_prior_TP15 \
 							--iter 15000 \
 							--mode histograms \
 							--num_points 10 \
@@ -10,7 +10,7 @@ python3 -m linfa.plot_disc --folder results/ \
 
 # python3 -m linfa.plot_disc --folder results/ \
 # 							--name test_lf_with_disc_hf_data_prior_TP15_rep_meas \
-# 							--iter 15000 \
+# 							--iter 2000 \
 # 							--mode discr_surface \
 # 							--num_points 40 \
 # 							--limfactor 2.0 \
