### 2023-12-12 eaf3167
update conditional statements to raise error on plot discrepancy surface
