### 2023-12-12 c863137
found bug in datageneration process with lambda two constant
