### 2023-12-12 1ba7d41
updated all test scripts with discrepancy to save surrogate to results folder
