+++ linfa/tests/test_plot_discr.sh
@@ -1,16 +1,19 @@
- # discr_surface
-# python3 -m linfa.plot_disc --folder results/ \
-# 							--name test_lf_with_disc_hf_data_prior_TP15 \
-# 							--iter 15000 \
-# 							--mode histograms \
-# 							--num_points 10 \
-# 							--limfactor 1.0 \
-
+# discr_surface
+python3 -m linfa.plot_disc --folder results/ \
+							--name test_lf_with_disc_hf_data_TP15_rep_meas \
+							--iter 2000 \
+							--mode histograms \
+							--num_points 10 \
+							--limfactor 1.0 \
 
 
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_lf_with_disc_hf_data_prior_TP15_rep_meas \
+							--name test_lf_with_disc_hf_data_TP15_rep_meas \
 							--iter 2000 \
 							--mode discr_surface \
 							--num_points 40 \
-							--limfactor 2.0 \
+							--limfactor 1.0 \
+
+python3 -m linfa.plot_res --folder results/ \
+							--name test_lf_with_disc_hf_data_TP15_rep_meas \
+							--iter 2000 \
