### 2023-12-13 512a690
adding dropouts to the mlp discrepancy and providing functions to interrogate the lf model plus discrepancy
