+++ linfa/tests/test_plot_discr.sh
@@ -1,19 +1,18 @@
 # discr_surface
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_lf_with_disc_hf_data_TP15_rep_meas \
-							--iter 2000 \
+							--name test_18_lf_w_disc_TP15_prior_rep_meas \
+							--iter 25000 \
 							--mode histograms \
 							--num_points 10 \
 							--limfactor 1.0 \
 
-
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_lf_with_disc_hf_data_TP15_rep_meas \
-							--iter 2000 \
+							--name test_18_lf_w_disc_TP15_prior_rep_meas \
+							--iter 25000 \
 							--mode discr_surface \
 							--num_points 40 \
 							--limfactor 1.0 \
 
 python3 -m linfa.plot_res --folder results/ \
-							--name test_lf_with_disc_hf_data_TP15_rep_meas \
-							--iter 2000 \
+							--name test_18_lf_w_disc_TP15_prior_rep_meas \
+							--iter 25000 \
