### 2023-12-15 4bd6c58
updated prior variance
