### 2023-12-15 3ef8380
updated prior variance
