+++ linfa/tests/test_discr_08_prior_TP1.py
@@ -165,7 +165,7 @@ def run_test():
         phys_inputs = transform.forward(calib_inputs)
         # Define prior moments
         pr_avg = torch.tensor([[1e3, -21e3]])
-        pr_std = torch.tensor([[1e3*0.01, 21e3*0.01]])
+        pr_std = torch.tensor([[1e3*0.1, 21e3*0.1]])
         # Eval log prior
         l1 = -0.5 * calib_inputs.size(1) * np.log(2.0 * np.pi)            
         l2 = (-0.5 * torch.log(torch.prod(pr_std))).item()
