### 2023-12-15 804b7bc
module to evaluate the model plus discrepancy at various TP and calib vars seems to work, need to call it with a script
