+++ linfa/tests/test_discr_06_LF_withdiscr_HFData_TP1.py
@@ -1,3 +1,4 @@
+from functools import partial
 from linfa.run_experiment import experiment
 from linfa.transform import Transformation
 from linfa.discrepancy import Discrepancy
