+++ linfa/tests/test_plot_discr.sh
@@ -1,18 +1,18 @@
-# discr_surface
-python3 -m linfa.plot_disc --folder results/ \
-							--name test_08_lf_w_disc_TP1_prior \
-							--iter 25000 \
-							--mode histograms \
-							--num_points 10 \
-							--limfactor 1.0 \
+# # discr_surface
+# python3 -m linfa.plot_disc --folder results/ \
+# 							--name test_08_lf_w_disc_TP1_prior \
+# 							--iter 25000 \
+# 							--mode histograms \
+# 							--num_points 10 \
+# 							--limfactor 1.0 \
 
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_08_lf_w_disc_TP1_prior \
+							--name test_09_lf_w_disc_TP1_prior \
 							--iter 25000 \
 							--mode discr_surface \
 							--num_points 40 \
 							--limfactor 1.0 \
 
-python3 -m linfa.plot_res --folder results/ \
-							--name test_08_lf_w_disc_TP1_prior \
-							--iter 25000 \
+# python3 -m linfa.plot_res --folder results/ \
+# 							--name test_08_lf_w_disc_TP1_prior \
+# 							--iter 25000 \
