### 2023-12-19 f0ec0cf
added training data with error bars to discrepancy plots
