+++ linfa/plot_disc.py
@@ -19,12 +19,12 @@ def scale_limits(min,max,factor):
 def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, out_dir, sample_size = 250):
 
     # Read result files
-    lf_model = np.loadtxt(lf_file)
-    lf_model_plus_disc = np.loadtxt(lf_dicr_file)
-    lf_model_plus_disc_plus_noise = np.loadtxt(lf_discr_noise_file)
-    data = np.loadtxt(data_file)
+    data                          = np.loadtxt(data_file)           # Experimental observations
+    lf_model                      = np.loadtxt(lf_file)             # Samples from low-fidelity posterior
+    lf_model_plus_disc            = np.loadtxt(lf_dicr_file)        # Samples from low-fidelity posterior + discrepancy
+    lf_model_plus_disc_plus_noise = np.loadtxt(lf_discr_noise_file) # Samples from low-fidelity posterior + discrepancy + noise
 
-    # Check for multiple TP-pairs
+    # Check for repated observations
     ## shape : no. var inputs pairs x no. batches
     num_dim = len(np.shape(lf_model))
     
@@ -32,13 +32,13 @@ def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, 
         # Plot histograms
         plt.figure(figsize = (6,4))
         ax = plt.gca()
-        plt.hist(lf_model,  label = r'$\eta \vert \mathbf{\theta}$', alpha = 0.5, density = True, hatch = '/')
-        plt.hist(lf_model_plus_disc,  label = r'$\zeta \vert \mathbf{\theta}, \delta$', alpha = 0.5, density = True)
-        plt.hist(lf_model_plus_disc_plus_noise,  label = r'$y \vert \mathbf{\theta}, \delta, \epsilon$', alpha = 0.5, density = True, hatch = '.')
+        plt.hist(lf_model, label = r'$\eta \vert \mathbf{\theta}$', alpha = 0.5, density = True, hatch = '/')
+        plt.hist(lf_model_plus_disc, label = r'$\zeta \vert \mathbf{\theta}, \delta$', alpha = 0.5, density = True)
+        plt.hist(lf_model_plus_disc_plus_noise, label = r'$y \vert \mathbf{\theta}, \delta, \epsilon$', alpha = 0.5, density = True, hatch = '.')
         plt.axvline(data[2], label = r'$y$', color = 'k', linewidth = 3)
         plt.legend(fontsize = 14)
         ax.set_xlabel('Coverage [ ]', fontweight = 'bold', fontsize = 16)
-        ax.set_ylabel('Density', fontweight = 'bold', fontsize = 16)
+        ax.set_ylabel('Density',      fontweight = 'bold', fontsize = 16)
         ax.xaxis.set_minor_locator(AutoMinorLocator())
         ax.yaxis.set_minor_locator(AutoMinorLocator())
         ax.xaxis.set_major_formatter(FormatStrFormatter("%.2f"))
@@ -58,14 +58,14 @@ def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, 
          ## Prepare for random sampling of batches
         lf_model_plus_disc = lf_model_plus_disc.reshape(len(temps), len(pressures), batch_size)
         random_array = np.random.randint(low = 0, high = batch_size, size = sample_size) # Randomly sample batch numbers without replacement
-        sample = np.zeros([len(temps), len(pressures), sample_size]) # Initialize
+        sample = np.zeros([len(temps), len(pressures), sample_size])                     # Initialize
 
          # For plotting
         plt.figure(figsize = (6,4))
         ax = plt.gca()
         clrs = ['b', 'm', 'r'] # Line colors for each temperature
[…]
