+++ linfa/tests/test_plot_discr.sh
@@ -7,7 +7,7 @@
 # 							--limfactor 1.0 \
 
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_09_lf_w_disc_TP1_prior \
+							--name test_18_lf_w_disc_TP15_prior_rep_meas \
 							--iter 25000 \
 							--mode discr_surface \
 							--num_points 40 \
