+++ linfa/tests/test_plot_discr.sh
@@ -6,10 +6,18 @@
 # 							--num_points 10 \
 # 							--limfactor 1.0 \
 
+# python3 -m linfa.plot_disc --folder results/ \
+# 							--name test_18_lf_w_disc_TP15_prior_rep_meas \
+# 							--iter 25000 \
+# 							--mode discr_surface \
+# 							--num_points 40 \
+# 							--limfactor 1.0 \
+
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_18_lf_w_disc_TP15_prior_rep_meas \
+							--name test_06_lf_w_disc_TP1 \
 							--iter 25000 \
-							--mode discr_surface \
+							--mode marginal_stats \
+							--saveinterval 1000 \
 							--num_points 40 \
 							--limfactor 1.0 \
 
