### 2023-12-20 cb0ea2c
added marginal posteriors
