+++ linfa/tests/test_plot_discr.sh
@@ -14,10 +14,9 @@
 # 							--limfactor 1.0 \
 
 python3 -m linfa.plot_disc --folder results/ \
-							--name test_06_lf_w_disc_TP1 \
+							--name test_18_lf_w_disc_TP15_prior_rep_meas \
 							--iter 25000 \
-							--mode marginal_stats \
-							--saveinterval 1000 \
+							--mode marginal_posterior \
 							--num_points 40 \
 							--limfactor 1.0 \
 
