+++ linfa/tests/test_plot_discr.sh
@@ -1,10 +1,10 @@
 # # discr_surface
-python3 -m linfa.plot_disc --folder results/ \
-							--name test_17_lf_w_disc_TP1_prior_rep_meas \
-							--iter 25000 \
-							--mode histograms \
-							--num_points 10 \
-							--limfactor 1.0 \
+# python3 -m linfa.plot_disc --folder results/ \
+# 							--name test_17_lf_w_disc_TP1_prior_rep_meas \
+# 							--iter 25000 \
+# 							--mode histograms \
+# 							--num_points 10 \
+# 							--limfactor 1.0 \
 
 # python3 -m linfa.plot_disc --folder results/ \
 # 							--name test_18_lf_w_disc_TP15_prior_rep_meas \
@@ -13,12 +13,11 @@ python3 -m linfa.plot_disc --folder results/ \
 # 							--num_points 40 \
 # 							--limfactor 1.0 \
 
-# python3 -m linfa.plot_disc --folder results/ \
-# 							--name test_17_lf_w_disc_TP1_prior_rep_meas \
-# 							--iter 25000 \
-# 							--mode marginal_posterior \
-# 							--num_points 40 \
-# 							--limfactor 1.0 \
+python3 -m linfa.plot_disc --folder results/ \
+							--name test_17_lf_w_disc_TP1_prior_rep_meas \
+							--iter 25000 \
+							--mode marginal_stats \
+							--saveinterval 1000 \
 
 # python3 -m linfa.plot_res --folder results/ \
 # 							--name test_08_lf_w_disc_TP1_prior \
