### 2023-12-21 7fe8345
started group analysis
