+++ linfa/plot_disc_groups.py
@@ -1,0 +1,213 @@
+import argparse
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.ticker import AutoMinorLocator, FormatStrFormatter, ScalarFormatter
+
+def get_data_labels(grouping):
+    if grouping == 'TP1' or 'TP15':
+        labels = ['none', 'prior', 'rep_meas', 'prior + rep_meas']
+    elif grouping == 'prior':
+        labels = ['TP1', 'TP15', 'TP1 + rep_meas', 'TP15 + rep_meas']
+    elif grouping == 'repeated measurements':
+        labels = ['TP1', 'TP15', 'TP1 + prior', 'TP15 + prior']
+    return labels
+
+def plot_marginal_stats(marg_stats_files, step_num, saveinterval, grouping, out_dirs):
+    
+    labels = get_data_labels(grouping)
+
+    # Get array of iterations where marginal statistics were saved
+    iterations = np.arange(start = saveinterval, 
+                           stop = step_num + saveinterval, 
+                           step = saveinterval, 
+                           dtype = int)
+    
+    # Initialize figure
+    fig, axes = plt.subplots(2, 2, figsize=(10, 5))
+    axes = axes.flatten()
+
+    color1 = '#1f78b4'  # Blue
+    color2 = '#33a02c'  # Teal
+    color3 = '#e31a1c'  # Red
+    color4 = '#ff7f00'  # Orange
+    colors = [color1, color2, color3, color4]
+    markers = ['v', 's', 'h','o']
+
+    # Loop over experiments
+    for loopA in range(len(marg_stats_files)):
+        
+        # Loop over save intervals
+        for loopB in range(len(iterations)):
+
+            # Read file
+            stats = np.loadtxt(marg_stats_files[loopA] + str(iterations[loopB])) 
+            
+            # Unpack sample statistics
+            mean = stats[:, 0]  # Sample mean
+            sd = stats[:, 1]    # Sample std. dev.
+
+            # Plot mean and sd of calibration parameters with adjusted data point label font size
+            axes[0].plot(iterations[loopB], mean[0], markers[loopA], color = colors[loopA], markersize = 8)
+            axes[1].plot(iterations[loopB], mean[1], markers[loopA], color = colors[loopA], markersize = 8)
+            axes[2].plot(iterations[loopB], sd[0], markers[loopA], color = colors[loopA], markersize = 8)
+            axes[3].plot(iterations[loopB], sd[1], markers[loopA], color = colors[loopA], markersize = 8)
+
+    # Set common labels for the y-axis
+    for ax, ylabel in zip(axes, [r'$\mathbb{E}(\theta_1 \vert \mathcal{D})$',
+                                 r'$\mathbb{E}(\theta_2 \vert \mathcal{D})$', 
+                                 r'$\mathbb{V}(\theta_1 \vert \mathcal{D})$', 
+                                 r'$\mathbb{V}(\theta_2 \vert \mathcal{D})$']):
+        ax.set_ylabel(ylabel, fontsize = 16, fontweight = 'bold')
+
+    # Add ground truth parameters to the sample average plots
+    gtCalParams = np.array([1.0E3, -21.0E3])
+    axes[0].axhline(gtCalParams[0], color = 'k', linewidth = 3, label = r'$\theta_1$')
+    axes[1].axhline(gtCalParams[1], color = 'k', linewidth = 3, label = r'$\theta_2$')
+    axes[0].legend(fontsize = 14)
+    axes[1].legend(fontsize = 14)
+
+    # Set x-axis label only for the bottom two subplots
+    for ax in axes[-2:]:
+        ax.set_xlabel('Iterations', fontsize = 16, fontweight = 'bold')
+
+    # Set tick label font size
[…]
