### 2023-12-26 d275d73
wrote code to plot marginal statistics vs. iterations in groups (e.g., TP1, TP15, etc.)
