+++ linfa/plot_disc_groups.py
@@ -2,20 +2,23 @@ import argparse
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.ticker import AutoMinorLocator, FormatStrFormatter, ScalarFormatter
+from matplotlib.lines import Line2D
 
-def get_data_labels(grouping):
-    if grouping == 'TP1' or 'TP15':
-        labels = ['none', 'prior', 'rep_meas', 'prior + rep_meas']
-    elif grouping == 'prior':
-        labels = ['TP1', 'TP15', 'TP1 + rep_meas', 'TP15 + rep_meas']
-    elif grouping == 'repeated measurements':
-        labels = ['TP1', 'TP15', 'TP1 + prior', 'TP15 + prior']
-    return labels
+# Set parameters common to all plots
+gtCalParams             = np.array([1.0E3, -21.0E3])
+colors                  = ['#1f78b4', '#33a02c', '#e31a1c', '#ff7f00'] # [Blue, Teal, Red, Orange]
+markers                 = ['v', 's', 'h', 'o']
+axis_label_font_size    = 16
+data_label_font_size    = 15
+legend_font_size        = 14
+marker_size             = 8
+line_width              = 3
+font_weight             = 'bold'
+dots_per_inch           = 300
 
-def plot_marginal_stats(marg_stats_files, step_num, saveinterval, grouping, out_dirs):
+def plot_marginal_stats(marg_stats_files, labels, step_num, saveinterval, grouping, out_dirs):
+    #  TODO: add location to savefile, add file format
     
-    labels = get_data_labels(grouping)
-
     # Get array of iterations where marginal statistics were saved
     iterations = np.arange(start = saveinterval, 
                            stop = step_num + saveinterval, 
@@ -23,63 +26,50 @@ def plot_marginal_stats(marg_stats_files, step_num, saveinterval, grouping, out_
                            dtype = int)
     
     # Initialize figure
-    fig, axes = plt.subplots(2, 2, figsize=(10, 5))
+    fig, axes = plt.subplots(2, 2, figsize = (14, 5))
     axes = axes.flatten()
 
-    color1 = '#1f78b4'  # Blue
-    color2 = '#33a02c'  # Teal
-    color3 = '#e31a1c'  # Red
-    color4 = '#ff7f00'  # Orange
-    colors = [color1, color2, color3, color4]
-    markers = ['v', 's', 'h','o']
-
-    # Loop over experiments
+    # Loop over experiments in group
     for loopA in range(len(marg_stats_files)):
         
         # Loop over save intervals
         for loopB in range(len(iterations)):
 
-            # Read file
-            stats = np.loadtxt(marg_stats_files[loopA] + str(iterations[loopB])) 
-            
-            # Unpack sample statistics
+            # Read file & unpack sample statistics
+            stats = np.loadtxt(marg_stats_files[loopA] + str(iterations[loopB]))
             mean = stats[:, 0]  # Sample mean
             sd = stats[:, 1]    # Sample std. dev.
 
-            # Plot mean and sd of calibration parameters with adjusted data point label font size
-            axes[0].plot(iterations[loopB], mean[0], markers[loopA], color = colors[loopA], markersize = 8)
-            axes[1].plot(iterations[loopB], mean[1], markers[loopA], color = colors[loopA], markersize = 8)
-            axes[2].plot(iterations[loopB], sd[0], markers[loopA], color = colors[loopA], markersize = 8)
-            axes[3].plot(iterations[loopB], sd[1], markers[loopA], color = colors[loopA], markersize = 8)
-
+            # Plot mean of calibration parameters
[…]
