### 2023-12-26 f765a52
changed lower bound of pre-exp factor, raised number of iterations
