### 2023-12-26 90ce624
lowered lower bound on first parameter, raised number of iterations
