### 2023-12-26 84299c8
increased standard deviation on parameter two in prior
