### 2023-12-26 716e6ab
increased no. of iterations, increased st dev in prior of second parameter, lowered lower bound of first parameter in tanh trsf
