### 2023-12-26 6331540
decreased std dev on parameter 2 in prior
