### 2023-12-26 a151e28
restored original prior values, raised lower bound on parameter 1 limits
