### 2023-12-28 316ed81
made all repeated observations 10, restored tranformation to original values
