+++ linfa/tests/test_discr_18_prior_TP15_rep_meas.py
@@ -197,7 +197,7 @@ def generate_data(use_true_model=False,num_observations=50):
 # Main code
 if __name__ == "__main__":
     
-    generate_data(use_true_model=True, num_observations=2)
+    generate_data(use_true_model = True, num_observations = 10)
     
     run_test()
 
