+++ linfa/tests/test_discr_12_LF_nodiscr_LFData_TP15_rep_meas.py
@@ -142,7 +142,7 @@ def generate_data():
     model = PhysChem(var_grid)
     
     # Generate data
-    model.genDataFile(use_true_model=False, num_observations = 2)
+    model.genDataFile(use_true_model=False, num_observations = 10)
 
 # Main code
 if __name__ == "__main__":
