+++ linfa/tests/test_discr_13_LF_nodiscr_HFData_TP1_rep_meas.py
@@ -137,7 +137,7 @@ def generate_data():
     model = PhysChem(var_grid)
     
     # Generate data
-    model.genDataFile(use_true_model=True,num_observations=2)
+    model.genDataFile(use_true_model = True, num_observations = 10)
 
 # Main code
 if __name__ == "__main__":
