+++ linfa/tests/test_discr_16_LF_withdiscr_HFData_TP15_rep_meas.py
@@ -177,7 +177,7 @@ def generate_data(use_true_model=False,num_observations=50):
 # Main code
 if __name__ == "__main__":
     
-    generate_data(use_true_model=True, num_observations=2)
+    generate_data(use_true_model = True, num_observations = 10)
     
     run_test()
 
