+++ linfa/tests/test_discr_15_LF_withdiscr_HFData_TP1_rep_meas.py
@@ -160,7 +160,7 @@ def run_test():
     # Run VI
     exp.run()
 
-def generate_data(use_true_model=False,num_observations=50):
+def generate_data(use_true_model=False,num_observations = 50):
 
     # Set variable grid
     var_grid = [[350.0],
@@ -175,7 +175,7 @@ def generate_data(use_true_model=False,num_observations=50):
 # Main code
 if __name__ == "__main__":
     
-    generate_data(use_true_model=True, num_observations=2)
+    generate_data(use_true_model = True, num_observations = 10)
     
     run_test()
 
