### 2023-12-29 7c6995f
Update test_linfa.py
Update RCR example with new annealing
