### 2024-01-05 e84f296
Update test_linfa.py
Pass annealing t into log_density function
