### 2024-01-09 #6108 @daneschi
Sorry to bother you again about this,
I just wanted to know if the paper was sent for review, if I need to
re-submit it or for anything else I can do on my end.
Thank you very much for your Kind Assistance with this!
with Best Regards,
Daniele

On Sat, Dec 9, 2023 at 3:05 PM Daniele Schiavazzi <
***@***.***> wrote:

> So sorry for the delay in getting back to you on this, I have missed your
> email.
> Possible reviewers are:
> - Alireza Doostan, University of Colorado, Boulder (
> ***@***.***)
> - Mingzhang Yin, University of Florida ***@***.***)
> - Bhattacharya Shrijita, Michigan State University ***@***.***)
> - Daniele Bigoni, Ammagamma s.r.l. ***@***.***)
> - Peng Chen, Georgia Tech ***@***.***)
> - Alessandro Veneziani, Emory University ***@***.***)
>
> On Sat, Dec 9, 2023 at 2:35 PM Arfon Smith ***@***.***>
> wrote:
>
>> @daneschi <https://github.com/daneschi> - thanks for your submission to
>> JOSS. We're currently managing a large backlog of submissions and the
>> editor most appropriate for your area is already rather busy.
>>
>> For now, we will need to waitlist this paper and process it as the queue
>> reduces. Thanks for your patience!
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/6108#issuecomment-1848643188>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/ADJL6EILU3T5YK2ITGUN6KLYIS4RLAVCNFSM6AAAAABADAOVQGVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMYTQNBYGY2DGMJYHA>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>
