### 2024-01-14 c20a347
updated plot_marginal_posterior to include ground truth values of parameters
