### 2024-01-14 40b19b0
added case study on linear trsf, found bug?
