### 2024-01-20 #6108 @arfon
> @assign me as editor

@lrnv - you need to do `@editorialbot assign me as editor`
