### 2024-01-20 #6108 @arfon
@daneschi – it looks like GitHub might have filtered out the email addresses of the people you listed [here](https://github.com/openjournals/joss-reviews/issues/6108#issuecomment-1848664121). Perhaps you could share them directly with @lrnv ?
