### 2024-01-20 #6108 @daneschi
Possible reviewers are:
- Alireza Doostan, University of Colorado, Boulder (
***@***.***)
- Mingzhang Yin, University of Florida ***@***.***)
- Bhattacharya Shrijita, Michigan State University ***@***.***)
- Daniele Bigoni, Ammagamma s.r.l. ***@***.***)
- Peng Chen, Georgia Tech ***@***.***)
- Alessandro Veneziani, Emory University ***@***.***)
Thanks!
Daniele

On Sat, Jan 20, 2024 at 9:25 AM The Open Journals editorial robot <
***@***.***> wrote:

> Assigned! @lrnv <https://github.com/lrnv> is now the editor
>
