### 2024-01-25 4894190
added dropouts to discrepancy
