### 2024-01-25 4b5c513
file set up to debug plot discrepancy surface function in plot_disc.py
