+++ linfa/tests/test_plot_discr.sh
@@ -1,18 +1,18 @@
-# # True process posterior
-python3 -m linfa.plot_disc --folder results/ \
-							--name test_06_lf_w_disc_TP1_linear \
-							--iter 5000 \
-							--mode histograms \
-							--num_points 10 \
-							--limfactor 1.0 \
-
-# # Discrepancy surface
+# True process posterior
 # python3 -m linfa.plot_disc --folder results/ \
-# 							--name test_06_lf_w_disc_TP1_linear \
-# 							--iter 5000 \
-# 							--mode discr_surface \
-# 							--num_points 40 \
+# 							--name test_07_lf_w_disc_TP15 \
+# 							--iter 25000 \
+# 							--mode histograms \
+# 							--num_points 10 \
 # 							--limfactor 1.0 \
+
+# Discrepancy surface
+python3 -m linfa.plot_disc --folder results/ \
+							--name test_07_lf_w_disc_TP15 \
+							--iter 25000 \
+							--mode discr_surface \
+							--num_points 40 \
+							--limfactor 1.0 \
 
 # # Marginal statistics
 # python3 -m linfa.plot_disc --folder results/ \
@@ -29,8 +29,8 @@ python3 -m linfa.plot_disc --folder results/ \
 
 # # Loss vs. iterations plot, calibration parameters posterior
 # python3 -m linfa.plot_res --folder results/ \
-# 						--name test_06_lf_w_disc_TP1_linear \
-# 						--iter 50000 \
+# 						--name test_07_lf_w_disc_TP15 \
+# 						--iter 25000 \
 
 # # Marginal statistics for groups to study
 # python3 -m linfa.plot_disc_groups --folder results/ \
