### 2024-01-25 92ba462
modified plot plot_discr_surface_2d function, cleaned up code, added print messages to compare train and test data
