+++ linfa/plot_disc.py
@@ -7,13 +7,13 @@ import argparse
 from numpy import random
 from matplotlib.ticker import AutoMinorLocator, FormatStrFormatter, ScalarFormatter
 
-def scale_limits(min,max,factor):
+def scale_limits(min, max, factor):
     if(min > max):
         temp = max
         max = min
         min = temp
-    center = 0.5 * (min + max)
     range = max - min
+    center = 0.5 * (max + min)
     return center - factor * 0.5 * range, center + factor * 0.5 * range
 
 def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, step_num, out_dir, img_format = 'png', sample_size = 250):
@@ -51,7 +51,6 @@ def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, 
         ax.xaxis.set_major_formatter(FormatStrFormatter("%.2f"))
         ax.tick_params(axis = 'both', which = 'both', direction = 'in', top = True, right = True, labelsize = 15)
         plt.tight_layout()
-        #fig.savefig('test.%s' % fmt, format=fmt)
         plt.savefig(out_dir + 'hist_' + str(step_num) +'.%s' % img_format, format = img_format, bbox_inches = 'tight', dpi = 300)
         plt.close()
    
@@ -117,6 +116,47 @@ def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, 
         plt.savefig(out_dir + 'hist_' + str(step_num) +'.%s' % img_format, format = img_format, bbox_inches = 'tight', dpi = 300)
         plt.close()
 
+def prep_test_grid(vars, limit_factor, no_grid_pts):
+    
+    ## Get min and max values of the variable inputs
+    # Variable input 1
+    min_dim_1 = torch.min(vars[:,0])
+    max_dim_1 = torch.max(vars[:,0])
+    min_dim_1, max_dim_1 = scale_limits(min_dim_1, max_dim_1, limit_factor)
+    
+    # Variable input 2
+    min_dim_2 = torch.min(vars[:,1])
+    max_dim_2 = torch.max(vars[:,1])
+    min_dim_2, max_dim_2 = scale_limits(min_dim_2, max_dim_2, limit_factor)
+
+    # Create test grid of variable inputs to evaluate discrepancy surrogate
+    test_grid_1 = torch.linspace(min_dim_1, max_dim_1, no_grid_pts)
+    test_grid_2 = torch.linspace(min_dim_2, max_dim_2, no_grid_pts)
+    grid_t, grid_p = torch.meshgrid(test_grid_1, test_grid_2, indexing='ij')
+    test_grid = torch.cat((grid_t.reshape(-1,1), grid_p.reshape(-1,1)),1)
+
+    return test_grid
+
+def print_disc_stats(avg_disc, disc_bounds):
+    print('\n')
+    print('Avg. Disc.\t LB \t\t UB')
+    print('----------------------------------------')
+    for loopA in range(len(disc_bounds[0])):
+        print('{:<10}\t{:<10}\t{:<10}'.format(round(avg_disc[loopA], 4), round(disc_bounds[0][loopA], 4), round(disc_bounds[1][loopA], 4)))
+        # Check for violation of bounds
+        if avg_disc[loopA] < disc_bounds[0][loopA] or avg_disc[loopA] > disc_bounds[1][loopA]:
+            print('Error in percentile calculation. Exiting...')
+            exit()
+
+def compare_disc(train, test):
+    print('\n')
+    print('Train Disc. \t Test Disc.')
+    print('----------------------------')
+    for loopA in range(len(train)):
[…]
