### 2024-01-25 1b56cc7
added docstrings and additional comments, pretty_print function for debugging, and TODO note on forward function to update test data statistics for normalizaiton
