+++ linfa/discrepancy.py
@@ -8,69 +8,97 @@ from linfa.mlp import FNN
 torch.set_default_tensor_type(torch.DoubleTensor)
 
 class Discrepancy(object):
-    """Class to create surrogate models for discrepancy estimatio
-       TO BE COMPLETED!!!
-    """
-    def __init__(self, model_name,
-                       lf_model,
-                       input_size,
-                       output_size,
-                       var_grid_in,
-                       var_grid_out,
-                       dnn_arch=None, 
-                       dnn_activation='relu', 
-                       model_folder='./', 
-                       surrogate=None, 
-                       device='cpu'):
+    '''Class to create surrogate models for discrepancy estimation.
+    
+    This class is designed to create surrogate models for estimating discrepancies 
+    in a model using a neural network (FNN). The discrepancy model is trained on a 
+    pre-defined grid of variable inputs and corresponding outputs.
+    
+    Args:
+        model_name (str):                   Name of the model.
+        lf_model (object):                  Low-fidelity model with a `solve_t` function.
+        input_size (int):                   Input size for FNN.
+        output_size (int):                  Output size for FNN.
+        var_grid_in (list of lists):        Variable grid locations for training.
+        var_grid_out (list):                Output variables.
+        dnn_arch (list of int, optional):   List containing the number of neurons for each hidden layer.
+        dnn_activation (str, optional):     Activation function for the neural network.
+        model_folder (str, optional):       Folder to save the model in.
+        surrogate (object, optional):       Pre-existing surrogate model.
+        device (str, optional):             Computational device.
+    
+    Attributes:
+        model_name (str):                   Name of the model.
+        model_folder (str):                 Folder to save the model in.
+        device (str):                       Computational device.
+        input_size (int):                   Input size for FNN.
+        output_size (int):                  Output size for FNN.
+        dnn_arch (list of int):             List containing the number of neurons for each hidden layer.
+        is_trained (boolean):               Flag for trained FNN.
+        lf_model (object):                  Low-fidelity model with a `solve_t` function.
+        var_grid_in (list of lists):        Store variable grid locations.
+        var_grid_out (list):                Output variables - multiple noisy observations are available for each combination of variables.
+        var_in_avg (torch.Tensor):          Average of variable inputs.
+        var_in_std (torch.Tensor):          Standard deviation of variable inputs.
+        var_out_avg (torch.Tensor):         Average of variable outputs.
+        var_out_std (torch.Tensor):         Standard deviation of variable outputs.
+        surrogate (object):                 Surrogate model (FNN).
 
-        self.model_name = model_name
-        self.model_folder = model_folder
+    Methods:
+        surrogate_save():                   Save the surrogate model to files.
+        surrogate_load():                   Load the surrogate model from files.
+        pretty_print():                     Print formatted information about the surrogate.
+        update():                           Train the surrogate model with a pre-grid.
+        forward():                          Evaluate the surrogate model.
+    '''
+    
+    def __init__(self, model_name, lf_model, input_size, output_size, var_grid_in, var_grid_out,
+                dnn_arch = None, dnn_activation = 'relu', model_folder = './', surrogate = None, 
+                device = 'cpu'):
+
+        self.model_name     = model_name        # (str): Name of the model
[…]
