### 2024-01-25 3d33685
incorporated discrepancy plots from des_debug_discr
