### 2024-01-29 #6108 @lrnv
:wave: @fabianhofmann & @mancellin & @davidtourigny, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
