### 2024-01-29 #6108 @lrnv
:wave: @Daniel-Dodd @vankesteren @robmoss, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
