### 2024-01-29 #6108 @davidtourigny
HelloSimilarly, I don’t have much expertise in this area and so can’t review, sorry :)Best wishesDavidOn Jan 29, 2024, at 12:28 PM, Erik-Jan van Kesteren ***@***.***> wrote:﻿
Sorry, have just accepted another JOSS review (just waiting on being assigned). Good luck!!
