### 2024-01-30 #6108 @robmoss
> 👋 @daniel-dodd @vankesteren @robmoss, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

I have limited familiarity with variational inference, but I can review this paper. This will be my first JOSS review, so I may have some questions about the review criteria.
