### 2024-01-31 #6108 @selimfirat
> 👋 @selimfirat, @ldliao, @atzberg, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)

Thanks for the invitation. I am interested in reviewing it.

I have some background knowledge on the topic of the paper, but this will be my first time reviewing for JOSS.
