### 2024-01-31 #6309 @editorialbot
[REVIEW]: LINFA: a Python library for variational inference with normalizing flow and annealing
**Submitting author:** @daneschi (<a href="http://orcid.org/0000-0001-9205-5989">Daniele E. Schiavazzi</a>)
**Repository:** https://github.com/desResLab/LINFA
**Branch with paper.md** (empty if default branch): master
**Version:** 1.5.1
