### 2024-02-01 desResLab/LINFA#7 @robmoss
linfa.models is not included in built packages
Hi @daneschi, I've started my [JOSS review](https://github.com/openjournals/joss-reviews/issues/6309) and encountered the following issue:

```sh
python3 -m venv venv
. venv/bin/activate
pip install linfa-vi
python -m unittest linfa.linfa_test_suite.trivial_example
# ModuleNotFoundError: No module named 'linfa.models'
```

I can see that `linfa.models` exists in the repository itself, and the tests run successfully on GitHub CI. The cause appears to be that `linfa.models` isn't being included in the wheel or sdist. However, because the repository has a ["flat layout"](https://packaging.python.org/en/latest/discussions/src-layout-vs-flat-layout/), `linfa.models` can be imported from the source directory when running the CI tests.

The packages can be fixed by making the following changes to `pyproject.toml`:

```toml
[tool.setuptools]
packages = ["linfa", "linfa.models", "linfa.tests"]
```

or. to avoid having to manually list each sub-package:

```toml
[tool.setuptools]
packages.find.include = ["linfa*"]
```

If you would like to ensure that the CI tests are being run against the built package, you could switch to a "src layout" (move `linfa` to `src/linfa`), and install the built package before running the tests. It's a little fiddlier, and a tool such as [tox](https://tox.wiki/en/latest/) or [nox](https://nox.thea.codes/en/stable/) can be helpful for this kind of thing. This is just a thought, I'm not suggesting you should do this as part of the JOSS review.
