### 2024-02-01 desResLab/LINFA#8 @daneschi
Sorry this is my first review with JOSS.
Should I fix these problems right away or wait until the review process is
concluded?
Thanks!

On Thu, Feb 1, 2024 at 3:04 AM Oskar Laverny ***@***.***>
wrote:

> Joss's editorialbot reports:
>
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - None
>
> MISSING DOIs
>
> - 10.1016/b978-0-08-051581-6.50057-x may be a valid DOI for title: Stochastic relaxation, Gibbs distributions, and the Bayesian restoration of images
> - 10.2172/4390578 may be a valid DOI for title: Equation of state calculations by fast computing machines
> - 10.1093/oso/9780198509936.003.0015 may be a valid DOI for title: Monte Carlo sampling methods using Markov chains and their applications
> - 10.21236/ada208388 may be a valid DOI for title: Sampling-based approaches to calculating marginal densities
> - 10.1561/9781601981851 may be a valid DOI for title: Graphical models, exponential families, and variational inference
> - 10.1109/tpami.2020.2992934 may be a valid DOI for title: Normalizing flows: An introduction and review of current methods
> - 10.1016/j.jcp.2022.111454 may be a valid DOI for title: Variational inference with NoFAS: Normalizing flow with adaptive surrogate for computationally expensive models
> - 10.1615/int.j.uncertaintyquantification.2022043110 may be a valid DOI for title: AdaAnn: Adaptive Annealing Scheduler for Probability Density Approximation
> - 10.1016/s0951-8320(02)00229-6 may be a valid DOI for title: Theorems and examples on high dimensional model representation
> - 10.1016/0041-5553(67)90144-9 may be a valid DOI for title: On the distribution of points in a cube and the approximate evaluation of integrals
>
> INVALID DOIs
>
> - None
>
> Please check for each of them that the proposed DOI is OK, and if so add
> it to your bib file.
>
