### 2024-02-01 893e986
fixed discrepancy plotting and now weights are init at every update
