+++ run_disc_examples.sh
@@ -1,0 +1,2 @@
+# Run case
+python -m unittest linfa.linfa_test_suite_discrepancy.test_discr_07_LF_withdiscr_HFData_TP15
