### 2024-02-06 e58d37d
updated sript to run larger bounds of tanh transformation. found bug in running multiple obervations in the plot discrepancy fxn
