+++ linfa/tests/test_discr_16_LF_withdiscr_HFData_TP15_rep_meas.py
@@ -24,7 +24,7 @@ def run_test():
     exp.input_size          = 2             # int: Dimensionalty of input (default 2)
     exp.batch_size          = 200           # int: Number of samples generated (default 100)
     exp.true_data_num       = 2             # double: Number of true model evaluted (default 2)
-    exp.n_iter              = 50000         # int: Number of iterations (default 25001)
+    exp.n_iter              = 25000         # int: Number of iterations (default 25001)
     exp.lr                  = 0.001         # float: Learning rate (default 0.003)
     exp.lr_decay            = 0.9999        # float:  Learning rate decay (default 0.9999)
     exp.log_interal         = 10            # int: How often to show loss stat (default 10)
@@ -53,8 +53,8 @@ def run_test():
     exp.device = torch.device('cuda:0' if torch.cuda.is_available() and not exp.no_cuda else 'cpu')
 
     # Define transformation
-    trsf_info = [['tanh', -7.0, 7.0, 500.0, 1500.0],
-                 ['tanh', -7.0, 7.0, -30000.0, -15000.0]]
+    trsf_info = [['tanh', -20.0, 20.0, 500.0, 1500.0],
+                 ['tanh', -20.0, 20.0, -30000.0, -15000.0]]
     trsf = Transformation(trsf_info)
     
     # Apply the transformation
@@ -66,6 +66,7 @@ def run_test():
 
     # Define model
     langmuir_model = PhysChem(variable_inputs)
+    
     # Assign as experiment model
     exp.model = langmuir_model
 
@@ -162,7 +163,7 @@ def run_test():
     # Run VI
     exp.run()
 
-def generate_data(use_true_model=False,num_observations=50):
+def generate_data(use_true_model=False, num_observations=50):
 
     # Set variable grid
     var_grid = [[350.0, 400.0, 450.0],
@@ -172,7 +173,7 @@ def generate_data(use_true_model=False,num_observations=50):
     model = PhysChem(var_grid)
     
     # Generate data
-    model.genDataFile(use_true_model=use_true_model,num_observations=num_observations)
+    model.genDataFile(use_true_model = use_true_model,num_observations=num_observations)
 
 # Main code
 if __name__ == "__main__":
