### 2024-02-06 86d4aa3
updated plot_discr_surface to generate test_grid with fxn eval
