### 2024-02-06 3c68cdb
generalized discrepancy surface plot function to repeated observations
