### 2024-02-08 0a7105c
first script with dropout layers
