### 2024-02-15 #6309 @robmoss
@lrnv thanks for checking in. I've made some progress, but have had to take some personal leave due to unforeseen circumstances. I will return to work next week, and continue my review. I identified [a packaging issue](https://github.com/desResLab/LINFA/issues/7), which is related to the paper but only applies to the software itself. Should I ask the authors to respond to this issue as part of the review process? Thanks in advance for your advice :)
