### 2024-02-15 #6309 @lrnv
@robmoss thanks for your reply. It is definitely expected from you to raise issues that apply to the software: the subject an the content of the review should be the software, the paper is just a side-effect IMHO. So if you found bugs or problems with the software, yes you can ask the authors to fix things as part of the review process. You can ask for new features if you think they would be logically inserted into the current project, or even for more detailled documentation, if you think it could be improved, etc. Sky (and reason) are the limits.
