### 2024-02-27 desResLab/LINFA#7 @robmoss
Hi, just checking in and wondering if you've had a look at this issue?
