### 2024-02-27 #6309 @selimfirat
> @robmoss I see your review already started, how is it going ?
>
> @selimfirat It's been two weeks did you have the time to take a look at this submission ? If you need more time, simply tell me I can setup automatic reminders for you if you want.

Hello again. I requested revisions by creating issues in the repository regarding the unchecked items: https://github.com/desResLab/LINFA/issues

Now, I am waiting until  the author makes neccessary changes. Is there any other action I should take?
