### 2024-02-27 #6309 @lrnv
@selimfirat Well done, this is great. 👍

Note that you are allowed and encouraged to also make comments and revisions requests about the code itself and not only the paper (this is the main point of JOSS reviews: to get reviewers oppinon and knowledge on the codebase). E.g., if you think the tests cases are not good enough, or if some implementation of a functionality looks clunky to you, or if you think something should be done in some other manner, like an API that does not look right, etc...

Sky is your limit.

Otherwise, everything looks good on my part :)
