### 2024-03-03 desResLab/LINFA#8 @lrnv
@daneschi re-reading myself I am not sure I was clear: all issues raised by the reviewers and myself are supposed to be solved **before** the paper gets accepted, so that we can check for completeness.
