### 2024-03-03 desResLab/LINFA#8 @daneschi
Thank you for the clarification. I'll do my best to address all the
comments this week. Thanks again for all your comments and suggestions!!

On Sun, Mar 3, 2024, 10:11 AM Oskar Laverny ***@***.***>
wrote:

> @daneschi <https://github.com/daneschi> re-reading myself I am not sure I
> was clear: all issues raised by the reviewers and myself are supposed to be
> solved *before* the paper gets accepted.
>
