### 2024-03-03 desResLab/LINFA#8 @lrnv
Take your time, do not harsh yourself you have all the time you want :)
