### 2024-03-06 c38fec5
added multiple evaluations of discrepancy, changed measurement noise to be iid
