+++ linfa/plot_disc.py
@@ -16,8 +16,76 @@ def scale_limits(min, max, factor):
     center = 0.5 * (max + min)
     return center - factor * 0.5 * range, center + factor * 0.5 * range
 
-def plot_disr_histograms(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, step_num, out_dir, img_format = 'png', sample_size = 250):
+def plot_disr_histograms_dropouts(lf_file, lf_dicr_file, lf_discr_noise_file, data_file, step_num, out_dir,  drop_out_evals = 0, img_format = 'png',sample_size = 50):
+    
+    # Read result files
+    data                          = np.loadtxt(data_file)           # Experimental observations
+    lf_model                      = np.loadtxt(lf_file)             # Samples from low-fidelity posterior
+    
+    batch_size = len(lf_model[0])
+    temps = np.unique(data[:, 0])
+    pressures = np.unique(data[:, 1])
+    no_reps = len(data[0,2:])
+    observed = data[:,2:].reshape(len(temps), len(pressures), no_reps)
+    
+    # For plotting
+    plt.figure(figsize = (6,4))
+    ax = plt.gca()
+    clrs = ['b', 'm', 'r'] # Line colors for each temperature
+    mkrs = ['v', 's', 'o'] # Line colors for each temperature
+    lines = []             # List to store Line2D objects for legend lines
+    sample = np.zeros([len(temps), len(pressures), sample_size]) # Initialize
+    
+    # Loop over stochastic discrepancy
+    for loopA in range(drop_out_evals):
+        
+        lf_model_plus_disc            = np.loadtxt(lf_dicr_file + '_' + str(loopA))        # Samples from low-fidelity posterior + discrepancy
+        lf_model_plus_disc_plus_noise = np.loadtxt(lf_discr_noise_file + '_' + str(loopA)) # Samples from low-fidelity posterior + discrepancy + noise
+        
+         ## Prepare for random sampling of batches
+        lf_model_plus_disc = lf_model_plus_disc.reshape(len(temps), len(pressures), batch_size)
+        random_array = np.random.randint(low = 0, high = batch_size, size = sample_size) # Randomly sample batch numbers without replacement
+        
+        # Loop over temperatures
+        for loopB, temp in enumerate(temps):
+            
+            # Loop over pressures
+            for loopC, pressure in enumerate(pressures):
 
+                # Evalute random sample of true process posterior
+                sample[loopB, loopC] = lf_model_plus_disc[loopB, loopC, random_array]
+
+            # Plot function & save line properties for legend
+            plt.plot(np.tile(pressures, (sample_size, 1)).transpose(), sample[loopB], linewidth=0.1, color = clrs[loopB])
+            
+            for loopD in range(no_reps):
+                line = plt.plot(np.tile(pressures, (sample_size, 1)).transpose(), 
+                                observed[loopB, :, loopD], 
+                                color = clrs[loopB], 
+                                marker = mkrs[loopB],
+                                markeredgecolor = 'k',
+                                markersize = 8,
+                                linestyle = '')[0]
+            lines.append(line)
+        
+    # Manually create the legend with custom linewidth
+    legend = plt.legend(lines, ['{} K'.format(round(temp)) for temp in temps], fontsize = 14)
+
+    # Set the linewidth for the legend lines
+    for line in legend.get_lines():
+        line.set_linewidth(2.0)  # Adjust the linewidth as needed
+        
+    ax.set_xlabel('Pressure [Pa]', fontsize = 16, fontweight = 'bold')
[…]

[…]
