### 2024-03-11 desResLab/LINFA#7 @daneschi
Thank you for your comments, I have fixed this issue as you suggested by adding "linfa.models" to the list of packages.  Thank you also for the interesting suggestion to change the layout and use additional tools. I'll take a look at tox/nox.
