+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
 requires-python = ">=3.7"
 
 [tool.setuptools]
-packages = ["linfa", "linfa.tests"]
+packages = ["linfa", "linfa.models", "linfa.tests"]
 
 [project.optional-dependencies]
 dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]
