### 2024-03-11 desResLab/LINFA#8 @daneschi
Thanks for pointing this out! When available, the missing DOIs for the bib references where provided.
