### 2024-03-11 aaccc3f
added missing DOIs to paper.bib
