+++ paper/paper.bib
@@ -5,7 +5,8 @@
   number={6},
   pages={721--741},
   year={1984},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TPAMI.1984.4767596}
 }
 @article{metropolis1953equation,
   title={Equation of state calculations by fast computing machines},
@@ -15,13 +16,15 @@
   number={6},
   pages={1087--1092},
   year={1953},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi={10.1063/1.1699114}
 }
 @article{hastings1970monte,
   title={{M}onte {C}arlo sampling methods using {M}arkov chains and their applications},
   author={Hastings, W Keith},
   year={1970},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/biomet/57.1.97}
 }
 @article{gelfand1990sampling,
   title={Sampling-based approaches to calculating marginal densities},
@@ -31,7 +34,8 @@
   number={410},
   pages={398--409},
   year={1990},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/01621459.1990.10476213}
 }
 @article{wainwright2008graphical,
   title={Graphical models, exponential families, and variational inference},
@@ -41,7 +45,8 @@
   number={1--2},
   pages={1--305},
   year={2008},
-  publisher={Now Publishers, Inc.}
+  publisher={Now Publishers, Inc.},
+  doi={10.1561/2200000001}
 }
 
 @book{villani2009optimal,
@@ -49,7 +54,8 @@
   author={Villani, C{\'e}dric and others},
   volume={338},
   year={2009},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-3-540-71050-9}
 }
 
 @article{kobyzev2020normalizing,
@@ -60,7 +66,8 @@
   number={11},
   pages={3964--3979},
   year={2020},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TPAMI.2020.2992934}
 }
 
 @article{papamakarios2021normalizing,
@@ -90,7 +97,8 @@
   volume={467},
   pages={111454},
   year={2022},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcp.2022.111454}
 }
 
 @article{cobian2023adaann,
@@ -99,7 +107,8 @@
   journal={International Journal for Uncertainty Quantification},
   volume={13},
   year={2023},
-  publisher={Begel House Inc.}
+  publisher={Begel House Inc.},
+  doi={10.1615/Int.J.UncertaintyQuantification.2022043110}
 }
 
 @article{dinh2016density,
@@ -172,7 +181,8 @@
   number={1},
   pages={1--67},
   year={1991},
-  publisher={Institute of Mathematical Statistics}
+  publisher={Institute of Mathematical Statistics},
+  doi={10.1214/aos/1176347963}
 }
 
 @article{gramacy2007tgp,
@@ -181,7 +191,8 @@
   journal={Journal of Statistical Software},
   volume={19},
   pages={1--46},
-  year={2007}
+  year={2007},
+  doi={10.18637/jss.v019.i09}
 }
 
 @article{sobol2003theorems,
@@ -192,7 +203,8 @@
   number={2},
   pages={187--193},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0951-8320(02)00229-6}
 }
 
 @article{brennan2020greedy,
@@ -205,11 +217,11 @@
 }
 
 @inproceedings{siahkoohi2021preconditioned,
[…]
