+++ linfa/run_experiment.py
@@ -171,14 +171,14 @@ class experiment:
                     i += 1
                 prev_i = i
 
-                if self.scheduler == 'AdaAnn':
+                if (self.scheduler == 'AdaAnn'):
                     z0 = nf.base_dist.sample([self.M])
                     zk, _ = nf(z0)
                     log_qk = self.model_logdensity(zk)
                     dt = self.tol / torch.sqrt(log_qk.var())
                     dt = dt.detach()# .numpy()
 
-                if self.scheduler == 'Linear':
+                if (self.scheduler == 'Linear'):
                     dt = self.linear_step
         else:
             loglist = []
@@ -214,78 +214,82 @@ class experiment:
         x0 = nf.base_dist.sample([self.batch_size])
         xk, sum_log_abs_det_jacobians = nf(x0)
 
+        # Check for last iteration
+        last_it = (iteration == self.n_iter)
+
         # generate and save samples evaluation
-        if sampling and iteration % self.save_interval == 0:
-            print('--- Saving results at iteration '+str(iteration))
-            x00 = nf.base_dist.sample([self.n_sample])
-            xkk, _ = nf(x00)
-            
-            # Save surrogate grid - there is no grid for discrepancy
-            if self.surrogate and (self.surrogate_type == 'surrogate'):
-                np.savetxt(self.output_dir + '/' + self.name + '_grid_' + str(iteration), self.surrogate.grid_record.clone().cpu().numpy(), newline="\n")
-            
-            # Save log profile
-            np.savetxt(self.output_dir + '/' + self.log_file, np.array(log), newline="\n")
-            
-            # Save normalized domain samples
-            np.savetxt(self.output_dir + '/' + self.name + '_samples_' + str(iteration), xkk.data.clone().cpu().numpy(), newline="\n")
-            
-            # Save samples in the original space
-            if self.transform:
-                xkk_samples = self.transform.forward(xkk).data.cpu().numpy()
-                np.savetxt(self.output_dir + '/' + self.name + '_params_' + str(iteration), xkk_samples, newline="\n")
-            else:
-                xkk_samples = xkk.data.cpu().numpy()
-                np.savetxt(self.output_dir + '/' + self.name + '_params_' + str(iteration), xkk_samples, newline="\n")
-            
-            # Save marginal statistics
-            np.savetxt(self.output_dir + '/' + self.name + '_marginal_stats_' + str(iteration), np.concatenate((xkk_samples.mean(axis=0).reshape(-1,1),xkk_samples.std(axis=0).reshape(-1,1)),axis=1), newline="\n")
-            
-            # Save log density at the same samples
-            np.savetxt(self.output_dir + '/' + self.name + '_logdensity_' + str(iteration), self.model_logdensity(xkk).data.cpu().numpy(), newline="\n")
-            
-            # Save model outputs at the samples - If a model is defined
-            if self.transform:
-                if(self.surrogate_type == 'surrogate'):
-                    # Define noise when we use NoFAS
-                    stds = torch.abs(self.model.defOut).to(self.device) * self.model.stdRatio
-                    o00 = torch.randn(x00.size(0), self.model.data.shape[0]).to(self.device)
-                    noise = o00*stds.repeat(o00.size(0),1)
-                    # Compute outputs
-                    if self.surrogate:
-                        np.savetxt(self.output_dir + '/' + self.name + '_outputs_' + str(iteration), (self.surrogate.forward(xkk) + noise).data.cpu().numpy(), newline="\n")
-                    else:
-                        np.savetxt(self.output_dir + '/' + self.name + '_outputs_' + str(iteration), (self.model.solve_t(self.transform.forward(xkk)) + noise).data.cpu().numpy(), newline="\n")
-                elif(self.surrogate_type == 'discrepancy'):
-                    # Define noise when we use NoFAS
[…]
