+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "linfa_vi"
-version = "1.3.6"
+version = "1.3.8"
 description = "A Python library for inference with normalizing flow and annealing"
 readme = "README.md"
 authors = [{ name = "resDesLab ", email = "daniele.schiavazzi@gmail.com" }]
@@ -19,8 +19,8 @@ classifiers = [
 keywords = ["variational inference", "normalizig flow", "adaptive annealing", "physics-based modeling"]
 dependencies = [
     "torch==1.13.1",
-    "numpy==1.21.6",
-    "matplotlib==3.5.3",
+    "numpy",
+    "matplotlib",
     'tomli; python_version < "3.11"',    
 ]
 requires-python = ">=3.7"
