+++ .github/workflows/test_publish_pypi.yml
@@ -42,11 +42,8 @@ jobs:
       with:
         password: ${{ secrets.PYPI_API_TOKEN }}
 
-    - uses: actions/checkout@master
     - uses: codecov/codecov-action@v4
       with:
-        fail_ci_if_error: true # optional (default = false)
-        files: ./coverage1.xml,./coverage2.xml # optional
         flags: unittests # optional
         name: codecov-umbrella # optional
         token: ${{ secrets.CODECOV_TOKEN }} # required
