### 2024-03-19 d19fd6c
changed documentation, added join() in plot_res
