+++ docs/content/num_examples.rst
@@ -4,6 +4,8 @@ Numerical Examples
 .. toctree::
    :maxdepth: 2
 
+   tutorial_2d
+   tutorial_3d
    trivial
    sobol
    rc
