+++ docs/content/tutorial_2d.rst
@@ -1,0 +1,4 @@
+TUTORIAL: Two-dimensional ballistic example
+===========================================
+
+.. include:: ../../tutorial/tutorial_linfa_2d.ipynb
