+++ paper/paper.bib
@@ -247,3 +247,51 @@
   publisher={Russian Academy of Sciences, Branch of Mathematical Sciences},
   doi={10.1016/0041-5553(67)90144-9}
 }
+
+@article{abril2023pymc,
+  title={Py{MC}: a modern, and comprehensive probabilistic programming framework in {P}ython},
+  author={Abril-Pla, Oriol and Andreani, Virgile and Carroll, Colin and Dong, Larry and Fonnesbeck, Christopher J and Kochurov, Maxim and Kumar, Ravin and Lao, Junpeng and Luhmann, Christian C and Martin, Osvaldo A and others},
+  journal={PeerJ Computer Science},
+  volume={9},
+  pages={e1516},
+  year={2023},
+  publisher={PeerJ Inc.}
+}
+
+@article{luttinen2016bayespy,
+  title={Bayespy: variational {B}ayesian inference in {P}ython},
+  author={Luttinen, Jaakko},
+  journal={The Journal of Machine Learning Research},
+  volume={17},
+  number={1},
+  pages={1419--1424},
+  year={2016},
+  publisher={JMLR. org}
+}
+
+@article{bingham2019pyro,
+  title={Pyro: {D}eep universal probabilistic programming},
+  author={Bingham, Eli and Chen, Jonathan P and Jankowiak, Martin and Obermeyer, Fritz and Pradhan, Neeraj and Karaletsos, Theofanis and Singh, Rohit and Szerlip, Paul and Horsfall, Paul and Goodman, Noah D},
+  journal={Journal of machine learning research},
+  volume={20},
+  number={28},
+  pages={1--6},
+  year={2019}
+}
+
+@article{huggins2023pyvbmc,
+  title = {Py{VBMC}: Efficient {B}ayesian inference in {P}ython},
+  author = {Bobby Huggins and Chengkun Li and Marlon Tobaben and Mikko J. Aarnos and Luigi Acerbi},
+  publisher = {The Open Journal},
+  journal = {Journal of Open Source Software},
+  url = {https://doi.org/10.21105/joss.05428},
+  doi = {10.21105/joss.05428},
+  year = {2023},
+  volume = {8},
+  number = {86},
+  pages = {5428}
+}
+
+
+
+
