+++ paper/paper.md
@@ -113,6 +113,18 @@ LINFA is designed as a general inference engine and allows the user to define cu
 
 6. **User-defined hyperparameters** - A complete list of hyperparameters with a description of their functionality can be found in the Appendix.
 
+## Software for variational inference
+
+Other Python packages were found to provide an implementation of variational inference and a number of additional features. An incomplete list of these packages is reported below.
+
+- [PyMc](https://www.pymc.io/projects/examples/en/latest/variational_inference/variational_api_quickstart.html)[@abril2023pymc]
+- [BayesPy](https://bayespy.org/) (with an accompanying paper, [BayesPy: Variational Bayesian Inference in Python](https://arxiv.org/abs/1410.0870))[@luttinen2016bayespy].
+- [Pyro](https://docs.pyro.ai/en/stable/inference.html) (with [some examples](http://pyro.ai/examples/svi_part_i.html))[@bingham2019pyro].
+- [PyVBMC](https://acerbilab.github.io/pyvbmc/) has an accompanying [JOSS article](https://joss.theoj.org/papers/10.21105/joss.05428)[@huggins2023pyvbmc].
+- Online notebooks (see this [example](https://predictivesciencelab.github.io/data-analytics-se/lecture28/hands-on-28.html)) which implement variational inference from scratch in `pytorch`.
+
+LINFA is based on normalizing flow transformations and therefore can infer non linear parameter dependence, provides the ability to adaptively train a surrogate model (NoFAS) which significantly reduces the computational cost for inference of parameters for expensive computational models. Finally, provides an adaptive annealing algorithm (AdaAnn) which autonomouly selects the appropriate annealing steps based on the current approximation of the posterior distribution. 
+
 # Numerical benchmarks
 
 We tested LINFA on multiple problems. These include inference on unimodal and multi-modal posterior distributions specified in closed form, ordinary differential models and dynamical systems with gradients directly computed through automatic differentiation in PyTorch, identifiable and non-identifiable physics-based models with fixed and adaptive surrogates, and high-dimensional statistical models. Some of the above tests are included with the library and systematically tested using GitHub Actions. A detailed discussion of these test cases is provided in the Appendix. To run the test type
