### 2024-03-19 a255c94
fixed paper references and codecov script
