### 2024-03-19 5801da1
added contributing guide, added depandance on codecov package
