+++ README.md
@@ -98,6 +98,10 @@ In addition you need to specify a list of options as discussed in the [documenta
 
 Two step-by-step tutorials ([tutorial 1](tutorial/tutorial_linfa_2d.ipynb) and [tutorial 2](tutorial/tutorial_linfa_3d.ipynb)) are also available which will guide you through the an inference problem for a ballistic simulation.
 
+### Contributing
+
+If you are interested in contributing to this project, plase take a look at the [contributed guide](CONTRIBUTING.md) provided with LINFA.
+
 ### Citation
 
 Did you use LINFA? Please cite our paper using:
