### 2024-03-19 #6309 @daneschi
Hi Oskar,
Just a quick update.
I have addressed all issues on GitHub.
I'm still working on integrating code coverage reports with GitHub Actions,
and  rendering python notebooks in sphinx to integrate the tutorial with
the online documentation.
Best,
Daniele

On Mon, Mar 11, 2024 at 5:15 AM Oskar Laverny ***@***.***>
wrote:

> @daneschi <https://github.com/daneschi> it has been two weeks. Would it
> be possible to have a status update on the issues that were raised ?
>
