### 2024-03-24 b77eae5
added tests with uniform priors
