### 2024-03-25 bea5ccd
added tutorials to sphinx documentation
