+++ docs/content/num_examples.rst
@@ -4,12 +4,11 @@ Numerical Examples
 .. toctree::
    :maxdepth: 2
 
-   tutorial_2d
-   tutorial_3d
+   tutorial/tutorial_linfa_2d
+   tutorial/tutorial_linfa_3d
    trivial
    sobol
    rc
    rcr
    friedman
    
-
