+++ docs/content/tutorial_3d.rst
@@ -1,4 +1,0 @@
-TUTORIAL: Three-dimensional ballistic example
-=============================================
-
-.. include:: ../../tutorial/tutorial_linfa_3d.ipynb
