+++ docs/content/tutorial_2d.rst
@@ -1,4 +1,0 @@
-TUTORIAL: Two-dimensional ballistic example
-===========================================
-
-.. include:: ../../tutorial/tutorial_linfa_2d.ipynb
