### 2024-03-26 c9c5939
added libraries for readthedocs
