### 2024-03-26 #6309 @daneschi
@lrnv Sure, please find a short report here below.
Please let me know for any further questions you may have.
Thank you again for your Kind Assistance with this review!

Q1 - Double check authors and affiliations (including ORCIDs)
A1 - The affiliations are correct. Only three authors have a ORCID number,
i.e.:

Daniele E.  Schiavazzi - 0000-0001-9205-5989
Jonathan Hauenstein - 0000-0002-9252-8210
Emma Cobian - 0000-0001-9872-6413

Q2 - Make a release of the software with the latest changes from the review
and post the version number here. This is the version that will be used in
the JOSS paper.
A2 - The latest version has all the changes. The version number is 1.5.1.
Should I create a new release on GitHub?

Q3 - Archive the release on Zenodo/figshare/etc and post the DOI here.
A3 - I have archived the reslease on Zenodo (
https://zenodo.org/records/10883597) with DOI: 10.5281/zenodo.10883597

Q4 - Make sure that the title and author list (including ORCIDs) in the
archive match those in the JOSS paper.
A4 - Yes, the two authors list match.

Q5 - Make sure that the license listed for the archive is the same as the
software license.
A5 - Yes, both use a MIT licence.

On Tue, Mar 26, 2024 at 3:02 PM Oskar Laverny ***@***.***>
wrote:

> @daneschi <https://github.com/daneschi> Could you please go through the
> todolist I've just geenrated for you, and make a report for me ?
>
