### 2024-03-26 #6309 @daneschi
OK, Thanks!
Best,
D.

On Tue, Mar 26, 2024 at 5:39 PM Oskar Laverny ***@***.***>
wrote:

> Should I create a new release on GitHub?
>
> No everything's alright like that, perfect, thansk you very much. That was
> a very detailed report 🤣 !
>
> I'll handle my part of the acceptation process first thing tomorow.
>
