### 2024-03-27 #6309 @lrnv
@daneschi The title of the zenodo archive does not match the title of the paper... There is a lacking "variational". Please fix one or the other.
