### 2024-03-27 #6309 @daneschi
@lrnv Sorry, the title is now consistent.
Best,
D.

On Wed, Mar 27, 2024 at 4:49 AM Oskar Laverny ***@***.***>
wrote:

> @daneschi <https://github.com/daneschi> The title of the zenodo archive
> does not match the title of the paper... There is a lacking "variational".
> Please fix one or the other.
>
