### 2024-03-27 #6309 @daneschi
Hi @lrnv, should I fix these issues?
Can I change the version when submitting to the repo? Do I need to also
update the zip file on Zenodo?
Thanks!
D.

On Wed, Mar 27, 2024 at 9:47 AM The Open Journals editorial robot <
***@***.***> wrote:

> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.1109/TPAMI.1984.4767596 is OK
> - 10.1063/1.1699114 is OK
> - 10.1093/biomet/57.1.97 is OK
> - 10.1080/01621459.1990.10476213 is OK
> - 10.1561/2200000001 is OK
> - 10.1007/978-3-540-71050-9 is OK
> - 10.1109/TPAMI.2020.2992934 is OK
> - 10.1016/j.jcp.2022.111454 is OK
> - 10.1615/Int.J.UncertaintyQuantification.2022043110 is OK
> - 10.1214/aos/1176347963 is OK
> - 10.18637/jss.v019.i09 is OK
> - 10.1016/S0951-8320(02)00229-6 is OK
> - 10.1016/j.jcp.2012.07.022 is OK
> - 10.1016/0041-5553(67)90144-9 is OK
> - 10.21105/joss.05428 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: Normalizing flows for probabilistic modeling and i...
> - No DOI given, and none found for title: Variational inference with normalizing flows
> - No DOI given, and none found for title: Density estimation using real NVP
> - No DOI given, and none found for title: Glow: Generative flow with invertible 1x1 convolut...
> - No DOI given, and none found for title: Masked Autoregressive Flow for Density Estimation
> - No DOI given, and none found for title: Improved variational inference with inverse autore...
> - No DOI given, and none found for title: MADE: Masked autoencoder for distribution estimati...
> - No DOI given, and none found for title: Batch normalization: Accelerating deep network tra...
> - No DOI given, and none found for title: Differentially Private Normalizing Flows for Densi...
> - No DOI given, and none found for title: Greedy inference with structure-exploiting lazy ma...
> - No DOI given, and none found for title: Preconditioned training of normalizing flows for v...
> - 10.7717/peerj-cs.1516 may be a valid DOI for title: PyMC: a modern, and comprehensive probabilistic pr...
> - No DOI given, and none found for title: Bayespy: variational Bayesian inference in Python
> - No DOI given, and none found for title: Pyro: Deep universal probabilistic programming
>
> INVALID DOIs
>
> - None
>
>
