### 2024-03-27 #6309 @lrnv
Yep you should try to fix those. Looks like a unicode character in a figure label and a \symbfit macro that is not working ocrrectly ?

If you change the paper only and not the content of the software, no need to update the version number IMHO. Otherwise, yes, the version number and the zenodo archive need to be updated too. Tell me what you decide.
