### 2024-03-27 678e7bc
added package to draft and replaced unicode character in figure def
