+++ paper/paper.bib
@@ -255,7 +255,8 @@
   volume={9},
   pages={e1516},
   year={2023},
-  publisher={PeerJ Inc.}
+  publisher={PeerJ Inc.},
+  doi={10.7717/peerj-cs.1516}
 }
 
 @article{luttinen2016bayespy,
