### 2024-03-27 #6309 @daneschi
OK, I added a new latex package to the paper and pushed an updated version.
For the unicode character, I tried to replace it but I used ":" in all
figure definitions, so why is it complaining only about "fig:trivial"?
Also, when I compile the paper with a github action, I don't get any
warning message. So I'm a bit confused.
Thanks.

On Wed, Mar 27, 2024 at 11:43 AM Oskar Laverny ***@***.***>
wrote:

> Yep you should try to fix those. Looks like a unicode character in a
> figure label and a \symbfit macro that is not working ocrrectly ?
>
> If you change the paper only and not the content of the software, no need
> to update the version number IMHO. Otherwise, yes, the version number and
> the zenodo archive need to be updated too. Tell me what you decide.
>
