### 2024-03-27 #6309 @lrnv
So :

- \symbfit ?
- fig:trivial is not unicode-problematic, its problematic because you defined it twice.
