### 2024-03-27 81c4beb
fixed label in figure and added bm with package in header
