+++ paper/paper.md
@@ -41,6 +41,7 @@ header-includes:
 - '\usepackage{booktabs}'
 - '\usepackage{caption}'
 - '\usepackage{unicode-math}'
+- '\usepackage{bm}'
 ---
 
 # Summary
@@ -151,23 +152,23 @@ NSF CAREER grant #1942662.
 
 ### Variational inference with normalizing flow
 
-Consider the problem of estimating (in a Bayesian sense) the parameters $\symbfit{z}\in\symbfit{\mathcal{Z}}$ of a physics-based or statistical model
+Consider the problem of estimating (in a Bayesian sense) the parameters $\bm{z}\in\bm{\mathcal{Z}}$ of a physics-based or statistical model
 $$
-\symbfit{x} = \symbfit{f}(\symbfit{z}) + \symbfit{\varepsilon},
+\bm{x} = \bm{f}(\bm{z}) + \bm{\varepsilon},
 $$
-from the observations $\symbfit{x}\in\symbfit{\mathcal{X}}$ and a known statistical characterization of the error $\symbfit{\varepsilon}$.
-We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \symbfit{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\symbfit{z}_{0})$ into a close approximation $q_{K}(\symbfit{z}_{K})$ of a desired target posterior density $p(\symbfit{z}|\symbfit{x})$. This transformation can be determined by composing $K$ bijections 
+from the observations $\bm{x}\in\bm{\mathcal{X}}$ and a known statistical characterization of the error $\bm{\varepsilon}$.
+We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \bm{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\bm{z}_{0})$ into a close approximation $q_{K}(\bm{z}_{K})$ of a desired target posterior density $p(\bm{z}|\bm{x})$. This transformation can be determined by composing $K$ bijections 
 $$
-\symbfit{z}_{K} = F(\symbfit{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\symbfit{z}_{0}),
+\bm{z}_{K} = F(\bm{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\bm{z}_{0}),
 $$
 and evaluating the transformed density through the change of variable formula (see @villani2009optimal).
 
-In the context of variational inference, we seek to determine an _optimal_ set of parameters $\symbfit{\lambda}\in\symbfit{\Lambda}$ so that $q_{K}(\symbfit{z}_{K})\approx p(\symbfit{z}|\symbfit{x})$. Given observations $\symbfit{x}\in\mathcal{\symbfit{X}}$, a likelihood function $l_{\symbfit{z}}(\symbfit{x})$ (informed by the distribution of the error $\symbfit{\varepsilon}$) and prior $p(\symbfit{z})$, a NF-based approximation $q_K(\symbfit{z})$ of the posterior distribution $p(\symbfit{z}|\symbfit{x})$ can be computed by maximizing the lower bound to the log marginal likelihood $\log p(\symbfit{x})$ (the so-called _evidence lower bound_ or ELBO), or, equivalently, by minimizing a _free energy bound_ (see, e.g., @rezende2015variational).
[…]
