### 2024-03-27 8b24bd9
removed unicode-math library
