+++ paper/paper.md
@@ -40,7 +40,7 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 header-includes: |
 - '\usepackage{booktabs}'
 - '\usepackage{caption}'
-- '\usepackage{unicode-math}'
+# - '\usepackage{unicode-math}'
 ---
 
 # Summary
