+++ paper/paper.md
@@ -28,8 +28,8 @@ authors:
     affiliation: 1
 
 affiliations:
- - name: Department of Applied and Computational Mathematics and Statistics, University of Notre Dame, Notre Dame, IN 46556, USA.
-   index: 1
+  - name: Department of Applied and Computational Mathematics and Statistics, University of Notre Dame, Notre Dame, IN 46556, USA.
+    index: 1
 date: 9 November 2023
 bibliography: paper.bib
 
@@ -38,9 +38,9 @@ bibliography: paper.bib
 aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
 aas-journal: Astrophysical Journal <- The name of the AAS journal.
 header-includes: |
-- '\usepackage{booktabs}'
-- '\usepackage{caption}'
-# - '\usepackage{unicode-math}'
+  - '\usepackage{booktabs}'
+  - '\usepackage{caption}'
+  - '\usepackage{bm}'
 ---
 
 # Summary
@@ -151,23 +151,23 @@ NSF CAREER grant #1942662.
 
 ### Variational inference with normalizing flow
 
-Consider the problem of estimating (in a Bayesian sense) the parameters $\symbf{z}\in\symbf{\mathcal{Z}}$ of a physics-based or statistical model
+Consider the problem of estimating (in a Bayesian sense) the parameters $\bm{z}\in\bm{\mathcal{Z}}$ of a physics-based or statistical model
 $$
-\symbf{x} = \symbf{f}(\symbf{z}) + \symbf{\varepsilon},
+\bm{x} = \bm{f}(\bm{z}) + \bm{\varepsilon},
 $$
-from the observations $\symbf{x}\in\symbf{\mathcal{X}}$ and a known statistical characterization of the error $\symbf{\varepsilon}$.
-We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \symbf{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\symbf{z}_{0})$ into a close approximation $q_{K}(\symbf{z}_{K})$ of a desired target posterior density $p(\symbf{z}|\symbf{x})$. This transformation can be determined by composing $K$ bijections 
+from the observations $\bm{x}\in\bm{\mathcal{X}}$ and a known statistical characterization of the error $\bm{\varepsilon}$.
+We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \bm{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\bm{z}_{0})$ into a close approximation $q_{K}(\bm{z}_{K})$ of a desired target posterior density $p(\bm{z}|\bm{x})$. This transformation can be determined by composing $K$ bijections 
 $$
-\symbf{z}_{K} = F(\symbf{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\symbf{z}_{0}),
+\bm{z}_{K} = F(\bm{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\bm{z}_{0}),
 $$
 and evaluating the transformed density through the change of variable formula (see @villani2009optimal).
 
[…]
