### 2024-03-28 #6309 @lrnv
Could you try using \mathbf instead ? or even \bm from the `bm` package ?
