+++ paper/paper.md
@@ -152,23 +152,23 @@ NSF CAREER grant #1942662.
 
 ### Variational inference with normalizing flow
 
-Consider the problem of estimating (in a Bayesian sense) the parameters $\mathbf{z}\in\mathbf{\mathcal{Z}}$ of a physics-based or statistical model
+Consider the problem of estimating (in a Bayesian sense) the parameters $\symbfit{z}\in\symbfit{\mathcal{Z}}$ of a physics-based or statistical model
 $$
-\mathbf{x} = \mathbf{f}(\mathbf{z}) + \mathbf{\varepsilon},
+\symbfit{x} = \symbfit{f}(\symbfit{z}) + \symbfit{\varepsilon},
 $$
-from the observations $\mathbf{x}\in\mathbf{\mathcal{X}}$ and a known statistical characterization of the error $\mathbf{\varepsilon}$.
-We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \mathbf{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\mathbf{z}_{0})$ into a close approximation $q_{K}(\mathbf{z}_{K})$ of a desired target posterior density $p(\mathbf{z}|\mathbf{x})$. This transformation can be determined by composing $K$ bijections 
+from the observations $\symbfit{x}\in\symbfit{\mathcal{X}}$ and a known statistical characterization of the error $\symbfit{\varepsilon}$.
+We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \symbfit{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\symbfit{z}_{0})$ into a close approximation $q_{K}(\symbfit{z}_{K})$ of a desired target posterior density $p(\symbfit{z}|\symbfit{x})$. This transformation can be determined by composing $K$ bijections 
 $$
-\mathbf{z}_{K} = F(\mathbf{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\mathbf{z}_{0}),
+\symbfit{z}_{K} = F(\symbfit{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\symbfit{z}_{0}),
 $$
 and evaluating the transformed density through the change of variable formula (see @villani2009optimal).
 
-In the context of variational inference, we seek to determine an _optimal_ set of parameters $\mathbf{\lambda}\in\mathbf{\Lambda}$ so that $q_{K}(\mathbf{z}_{K})\approx p(\mathbf{z}|\mathbf{x})$. Given observations $\mathbf{x}\in\mathcal{\mathbf{X}}$, a likelihood function $l_{\mathbf{z}}(\mathbf{x})$ (informed by the distribution of the error $\mathbf{\varepsilon}$) and prior $p(\mathbf{z})$, a NF-based approximation $q_K(\mathbf{z})$ of the posterior distribution $p(\mathbf{z}|\mathbf{x})$ can be computed by maximizing the lower bound to the log marginal likelihood $\log p(\mathbf{x})$ (the so-called _evidence lower bound_ or ELBO), or, equivalently, by minimizing a _free energy bound_ (see, e.g., @rezende2015variational).
[…]
