+++ paper/paper.md
@@ -28,8 +28,8 @@ authors:
     affiliation: 1
 
 affiliations:
-  - name: Department of Applied and Computational Mathematics and Statistics, University of Notre Dame, Notre Dame, IN 46556, USA.
-    index: 1
+ - name: Department of Applied and Computational Mathematics and Statistics, University of Notre Dame, Notre Dame, IN 46556, USA.
+   index: 1
 date: 9 November 2023
 bibliography: paper.bib
 
@@ -38,9 +38,9 @@ bibliography: paper.bib
 aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
 aas-journal: Astrophysical Journal <- The name of the AAS journal.
 header-includes: |
-  - '\usepackage{booktabs}'
-  - '\usepackage{caption}'
-  - '\usepackage{bm}'
+- '\usepackage{booktabs}'
+- '\usepackage{caption}'
+- '\usepackage{unicode-math}'
 ---
 
 # Summary
@@ -151,23 +151,23 @@ NSF CAREER grant #1942662.
 
 ### Variational inference with normalizing flow
 
-Consider the problem of estimating (in a Bayesian sense) the parameters $\symbfit{z}\in\symbfit{\mathcal{Z}}$ of a physics-based or statistical model
+Consider the problem of estimating (in a Bayesian sense) the parameters $\symbf{z}\in\symbf{\mathcal{Z}}$ of a physics-based or statistical model
 $$
-\symbfit{x} = \symbfit{f}(\symbfit{z}) + \symbfit{\varepsilon},
+\symbf{x} = \symbf{f}(\symbf{z}) + \symbf{\varepsilon},
 $$
-from the observations $\symbfit{x}\in\symbfit{\mathcal{X}}$ and a known statistical characterization of the error $\symbfit{\varepsilon}$.
-We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \symbfit{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\symbfit{z}_{0})$ into a close approximation $q_{K}(\symbfit{z}_{K})$ of a desired target posterior density $p(\symbfit{z}|\symbfit{x})$. This transformation can be determined by composing $K$ bijections 
+from the observations $\symbf{x}\in\symbf{\mathcal{X}}$ and a known statistical characterization of the error $\symbf{\varepsilon}$.
+We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \symbf{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\symbf{z}_{0})$ into a close approximation $q_{K}(\symbf{z}_{K})$ of a desired target posterior density $p(\symbf{z}|\symbf{x})$. This transformation can be determined by composing $K$ bijections 
 $$
-\symbfit{z}_{K} = F(\symbfit{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\symbfit{z}_{0}),
+\symbf{z}_{K} = F(\symbf{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\symbf{z}_{0}),
 $$
 and evaluating the transformed density through the change of variable formula (see @villani2009optimal).
 
[…]
