### 2024-03-28 a3beb4c
changed to linear transformation and fixed support on uniform prior
