### 2024-03-28 60361b6
corrected bug in uniform prior
