+++ paper/paper.md
@@ -37,7 +37,7 @@ bibliography: paper.bib
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
 aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
 aas-journal: Astrophysical Journal <- The name of the AAS journal.
-header-includes: |
+header-includes:
 - '\usepackage{booktabs}'
 - '\usepackage{caption}'
 ---
@@ -150,23 +150,23 @@ NSF CAREER grant #1942662.
 
 ### Variational inference with normalizing flow
 
-Consider the problem of estimating (in a Bayesian sense) the parameters $\symbfit{z}\in\symbfit{\mathcal{Z}}$ of a physics-based or statistical model
+Consider the problem of estimating (in a Bayesian sense) the parameters $\boldsymbol{z}\in\boldsymbol{\mathcal{Z}}$ of a physics-based or statistical model
 $$
-\symbfit{x} = \symbfit{f}(\symbfit{z}) + \symbfit{\varepsilon},
+\boldsymbol{x} = \boldsymbol{f}(\boldsymbol{z}) + \boldsymbol{\varepsilon},
 $$
-from the observations $\symbfit{x}\in\symbfit{\mathcal{X}}$ and a known statistical characterization of the error $\symbfit{\varepsilon}$.
-We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \symbfit{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\symbfit{z}_{0})$ into a close approximation $q_{K}(\symbfit{z}_{K})$ of a desired target posterior density $p(\symbfit{z}|\symbfit{x})$. This transformation can be determined by composing $K$ bijections 
+from the observations $\boldsymbol{x}\in\boldsymbol{\mathcal{X}}$ and a known statistical characterization of the error $\boldsymbol{\varepsilon}$.
+We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \boldsymbol{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\boldsymbol{z}_{0})$ into a close approximation $q_{K}(\boldsymbol{z}_{K})$ of a desired target posterior density $p(\boldsymbol{z}|\boldsymbol{x})$. This transformation can be determined by composing $K$ bijections 
 $$
-\symbfit{z}_{K} = F(\symbfit{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\symbfit{z}_{0}),
+\boldsymbol{z}_{K} = F(\boldsymbol{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\boldsymbol{z}_{0}),
 $$
 and evaluating the transformed density through the change of variable formula (see @villani2009optimal).
 
[…]
