+++ paper/paper.md
@@ -40,6 +40,7 @@ aas-journal: Astrophysical Journal <- The name of the AAS journal.
 header-includes:
 - '\usepackage{booktabs}'
 - '\usepackage{caption}'
+- '\usepackage{bm}'
 ---
 
 # Summary
@@ -150,23 +151,23 @@ NSF CAREER grant #1942662.
 
 ### Variational inference with normalizing flow
 
-Consider the problem of estimating (in a Bayesian sense) the parameters $\boldsymbol{z}\in\boldsymbol{\mathcal{Z}}$ of a physics-based or statistical model
+Consider the problem of estimating (in a Bayesian sense) the parameters $\bm{z}\in\bm{\mathcal{Z}}$ of a physics-based or statistical model
 $$
-\boldsymbol{x} = \boldsymbol{f}(\boldsymbol{z}) + \boldsymbol{\varepsilon},
+\bm{x} = \bm{f}(\bm{z}) + \bm{\varepsilon},
 $$
-from the observations $\boldsymbol{x}\in\boldsymbol{\mathcal{X}}$ and a known statistical characterization of the error $\boldsymbol{\varepsilon}$.
-We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \boldsymbol{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\boldsymbol{z}_{0})$ into a close approximation $q_{K}(\boldsymbol{z}_{K})$ of a desired target posterior density $p(\boldsymbol{z}|\boldsymbol{x})$. This transformation can be determined by composing $K$ bijections 
+from the observations $\bm{x}\in\bm{\mathcal{X}}$ and a known statistical characterization of the error $\bm{\varepsilon}$.
+We tackle this problem with variational inference and normalizing flow. A normalizing flow (NF) is a nonlinear transformation $F:\mathbb{R}^{d}\times \bm{\Lambda} \to \mathbb{R}^{d}$ designed to map an easy-to-sample \emph{base} distribution $q_{0}(\bm{z}_{0})$ into a close approximation $q_{K}(\bm{z}_{K})$ of a desired target posterior density $p(\bm{z}|\bm{x})$. This transformation can be determined by composing $K$ bijections 
 $$
-\boldsymbol{z}_{K} = F(\boldsymbol{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\boldsymbol{z}_{0}),
+\bm{z}_{K} = F(\bm{z}_{0}) = F_{K} \circ F_{K-1} \circ \cdots \circ F_{k} \circ \cdots \circ F_{1}(\bm{z}_{0}),
 $$
 and evaluating the transformed density through the change of variable formula (see @villani2009optimal).
 
[…]
