### 2024-04-04 #6309 @arfon
:wave: folks. A few things with the paper here: It looks to be 'off' in a few places:

- The AAS link in the margin - is this paper being jointly published with AAS? If not, please remove those sections from your `paper.md`.
- A table in the appendix is overflowing in the references. Please see if you can fix this with a `\newpage` or similar in the paper.
- The paper is *waaaaaaay* longer than we like to typically publish in JOSS (guideline is ~1000 words). Given this submission is ready to go I'm not going ask for changes now, but in the future @lrnv please look out for this.
