### 2024-04-04 #6309 @daneschi
Hi @lrnv should I fix these problems in the md file?
Sorry for the length, I also thought it was OK to have more material in the
appendix.
Thanks

On Thu, Apr 4, 2024 at 11:03 AM Oskar Laverny ***@***.***>
wrote:

> Thanks a lot @arfon <https://github.com/arfon> for catching theses. I
> will be more careful in the future w.r.t. text length. Since here ~75% of
> it is in appendix, I thought it was OK, but reading the guidelines once
> more, I see that it is not.
>
