### 2024-04-04 #6309 @lrnv
Yes please. You are not publishing in the AAS ? Thus the AAS section of your .md file should not be there.
