+++ paper/paper.md
@@ -35,8 +35,8 @@ bibliography: paper.bib
 
 # Optional fields if submitting to a AAS journal too, see this blog post:
 # https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
-aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
-aas-journal: Astrophysical Journal <- The name of the AAS journal.
+# aas-doi: 10.3847/xxxxx <- update this with the DOI from AAS once you know it.
+# aas-journal: Astrophysical Journal <- The name of the AAS journal.
 header-includes:
 - '\usepackage{booktabs}'
 - '\usepackage{caption}'
@@ -443,5 +443,7 @@ This section contains the list of all hyperparameters in the library, their defa
 \bottomrule
 \end{tabular}}
 \end{table}
+
+\newpage
 
 # References
