### 2024-04-05 #6309 @daneschi
Thank you all very much for your Kind Assistance with this review!
Thanks @robmoss, @selimfirat for your comments and @lrnv for your editorial
supervision and your patience!

On Fri, Apr 5, 2024 at 4:20 PM The Open Journals editorial robot <
***@***.***> wrote:

> 🎉🎉🎉 Congratulations on your paper acceptance! 🎉🎉🎉
>
> If you would like to include a link to your paper from your README use the
> following code snippets:
>
> Markdown:
>
> HTML:
> <a style="border-width:0" href="https://doi.org/10.21105/joss.06309">
> </a>
>
> reStructuredText:
>    :target: https://doi.org/10.21105/joss.06309
>
> This is how it will look in your documentation:
>
> [image: DOI] <https://doi.org/10.21105/joss.06309>
>
> *We need your help!*
>
> The Journal of Open Source Software is a community-run journal and relies
> upon volunteer effort. If you'd like to support us please consider doing
> either one (or both) of the the following:
>
>    - Volunteering to review for us sometime in the future. You can add
>    your name to the reviewer list here:
>    https://reviewers.joss.theoj.org/join
>    - Making a small donation to support our running costs here:
>    https://numfocus.org/donate-to-joss
>
